==> 1-gila-preprocessing_new.Rmd <==
---
title: Upper Gila Watershed Preprocessing
author: Lauren Victoria Jaramillo, UNM (alwayslo@unm.edu)
date: 11/29/2018
update: 8/21/2019
output: html_notebook
---

#Conditional Probabilities and Evidence#
This R notebook is used to process the existing hydrology (Gila at Gila River USGS gage No.09430500). Conditional probabilities for inundation and recession rates are calculated in this script. Recession rates were calculated using information for site 1 only for this recreation of the historical model/process. USGS data was used to generate conditional probabilities and scenario data was used to generate evidence.

This first code chunck clears the environment, installs all the necessary packages, and returns the current working directory to ensure filespaths will execute properly.
```{r}
rm(list=ls())
library("xts")
library("prob")
library("TTR")
library("bigmemory")
library("zoo")
library("raster")
library("sp")
library("rgdal")
library("rasterVis")
library("ggplot2")
library("rgdal")

getwd()
```

#### Process USGS hydrology data ####
The hydrologic historical data for the Gila River is located in usgs_09430500_daily.txt. Units are cfs. This code chunk loads usgs_09430500_daily.txt and uses it to create a dataframe called usgs_hydro.
```{r}
usgs_hydro <- read.csv("data/usgs_09430500_daily.txt", skip=28,header = TRUE, sep = "")
usgs_hydro <- setNames(usgs_hydro,c("agency","site_no","date","flowrate (cfs)","code"))
usgs_hydro$date<-as.Date(usgs_hydro$date, format="%Y-%m-%d")
```

#### Process Q-model bins ####
Establishes model q-bins and calculates the probability of discharge range for any given entire year and discrete time state (April-May, June-July, August-September).Model Q-bins were origninal selected so that inundation areas increased by less than 15% between the discharge associated with each bin (i.e. 1000 cfs, 1500 cfs, 2000 cfs, 2500 cfs, 3000 cfs, 3500 cfs, 4000 cfs). The number of bins were then increased based on a frequency analysis adding 8 more bins for a total of 15 bins (i.e.200 cfs,225 cfs,250 cfs,275 cfs,300 cfs,350 cfs,400 cfs,500 cfs,1000 cfs, 1500 cfs, 2000 cfs, 2500 cfs, 3000 cfs, 3500 cfs, 4000 cfs)

```{r}
q_bin <- c(200,225,250,275,300,350,400,500,1000, 1500, 2000, 2500, 3000, 3500, 4000)

```


#### Process recession rate bins ####
Establishes model recession rate bins and calculates the probability of recession rate ranges for 14-day and 30-day recession rates for any given entire year and discrete time state (April-May, June-July, August-September).

Recession rate bins were selected so that .....???? (i.e. 0 cm/day, 1 cm/day, 3 cm/day, 6 cm/day).
```{r}
recessionrate_bin <- c(0,1,3,6)
```

#### Process CUFA hydrology data ####
The hydrologic historical data for the Gila River is located in scenario_revised.csv and includes the 1-natural, 2-Alt. E, 3-Alt.B , and 4-Alt.A (no action)). This code chunk loads scenario_revised.csv (excluding the flow columns for downstream at Riverside) and uses it to create a dataframe called hydro.

```{r}
hydro <- read.csv("data/scenarios_revised.csv", skip=0,header = TRUE)[ ,1:5]
hydro <- setNames(hydro, c("date","natural", "Alt_E","Alt_B","Alt_A"))
hydro$date<-as.Date(hydro$date, format="%m/%d/%Y")
hydro <- cbind(hydro[1],usgs_hydro[4],hydro[2:4])
hydro <- setNames(hydro, c("date","natural", "Alt_E","Alt_B","Alt_A"))
```


###TEST####
# ```{r}
# hydro <- cbind(hydro[1:2],hydro[3]/2,hydro[4]/1,hydro[5])
# ```

This project includes three study sites. This code chunck adds the stage (both ft and cm) for each site based on the flowrate and previously developed regression equations for each site (1-North,2- Middle,3- South). The hydro column called must correspond with each scenario (i.e.1-natural, 2-Alt. E, 3-Alt.B , and 4-Alt.A (no action)).

North, Depth = 0.0003Q + 2
Middle, Depth = 0.0004Q + 2.1
South, Depth = 0.0003Q + 1.2


SPECIFY SITE

<!-- ###Site 1### -->
<!-- ```{r} -->
<!-- head_conv <- data.frame(hydro$date,0.0003*hydro[2]+2, 0.0003*hydro[3]+2,0.0003*hydro[4]+2,0.0003*hydro[5]+2) -->
<!-- head_conv <- setNames(head_conv, c("date","natural(ft)", "Alt_E(ft)","Alt_B(ft)","Alt_A(ft)")) -->

<!-- head_cm <- data.frame(hydro$date,head_conv[2]*30.48,head_conv[3]*30.48,head_conv[4]*30.48,head_conv[5]*30.48) -->
<!-- head_cm <- setNames(head_cm,c("date","natural(cm)", "Alt_E(cm)","Alt_B(cm)","Alt_A(cm)")) -->
<!-- ``` -->


###Site 2###
```{r}
head_conv <- data.frame(hydro$date,0.0004*hydro[2]+2.1, 0.0004*hydro[3]+2.1,0.0004*hydro[4]+2.1,0.0004*hydro[5]+2.1)
head_conv <- setNames(head_conv, c("date","natural(ft)", "Alt_E(ft)","Alt_B(ft)","Alt_A(ft)"))

head_cm <- data.frame(hydro$date,head_conv[2]*30.48,head_conv[3]*30.48,head_conv[4]*30.48,head_conv[5]*30.48)
head_cm <- setNames(head_cm,c("date","natural(cm)", "Alt_E(cm)","Alt_B(cm)","Alt_A(cm)"))
```

<!-- ###Site 3### -->
<!-- ```{r} -->
<!-- head_conv <- data.frame(hydro[1],0.0003*hydro[2]+1.2, 0.0003*hydro[3]+1.2,0.0003*hydro[4]+1.2,0.0003*hydro[5]+1.2) -->
<!-- head_conv <- setNames(head_conv, c("date","natural(ft)", "Alt_E(ft)","Alt_B(ft)","Alt_A(ft)")) -->

<!-- head_cm <- data.frame(hydro$date,head_conv[2]*30.48,head_conv[3]*30.48,head_conv[4]*30.48,head_conv[5]*30.48) -->
<!-- head_cm <- setNames(head_cm,c("date","natural(cm)", "Alt_E(cm)","Alt_B(cm)","Alt_A(cm)")) -->
<!-- ``` -->

#### Calculate recession rates ####
Calculates the 14-day recession rate (cm/day). Recession rate (forward) was calculated using a 14 day average based on the site specified above). The equation was mimicing that of Morrison & Stone (2014).
```{r}
# average_natural_stage_14_day <-sum(head_cm[1:14,1])/14
# recession_rate_natural_14_day <- (head_cm[1,1] - head_cm[14,1])/14

recessionrate_14d<- data.frame()

for (i in 1:(nrow(head_cm))) 
  {recessionrate_14d <- rbind(recessionrate_14d,c((head_cm[i,2]-head_cm[13+i,2])/14,
                                                  (head_cm[i,3]-head_cm[13+i,3])/14,
                                                  (head_cm[i,4]-head_cm[13+i,4])/14,
                                                  (head_cm[i,5]-head_cm[13+i,5])/14))}

recessionrate_14d <- cbind(hydro[,1],recessionrate_14d)
recessionrate_14d <- setNames(recessionrate_14d,c("date",
                                                  "natural(cm/d)",
                                                  "Alt_E(cm/d)",
                                                  "Alt_B(cm/d)",
                                                  "Alt_A(cm/d)"))
recessionrate_14d<-na.omit(recessionrate_14d)
```

Bins the 14-day recession rates for all scenarios and calculates the probability of the recession rate range for any given year.(Includes entire dataset for each scenario. No trimming of the data to exclude flows less than 1000 cfs because the recessionrate doesn't include any information about the magnitude of the flowrate).
```{r}
a <- recessionrate_14d[2:5]

rr_14d_bin <-cbind( recessionrate_14d[1], 
              recessionrate_bin_1 <- a <= recessionrate_bin[1],
              recessionrate_bin_2 <- recessionrate_bin[1]< a & a <=recessionrate_bin[2],
              recessionrate_bin_3 <- recessionrate_bin[2]< a & a <=recessionrate_bin[3],
              recessionrate_bin_4 <- recessionrate_bin[3]< a & a <=recessionrate_bin[4],
              recessionrate_bin_5 <- a > recessionrate_bin[4])

colnames(rr_14d_bin) <- c("date",
                          "rr_bin_1_natural","rr_bin_1_scenario1","rr_bin_1_scenario2","rr_bin_1_scenario3",
                          "rr_bin_2_natural","rr_bin_2_scenario1","rr_bin_2_scenario2","rr_bin_2_scenario3",
                          "rr_bin_3_natural","rr_bin_3_scenario1","rr_bin_3_scenario2","rr_bin_3_scenario3",
                          "rr_bin_4_natural","rr_bin_4_scenario1","rr_bin_4_scenario2","rr_bin_4_scenario3",
                          "rr_bin_5_natural","rr_bin_5_scenario1","rr_bin_5_scenario2","rr_bin_5_scenario3")

rr_14d_prob_yr<-data.frame()
rr_14d_prob_yr <-cbind(length(which(rr_14d_bin$rr_bin_1_natural))/length(rr_14d_bin$rr_bin_1_natural),
                                       length(which(rr_14d_bin$rr_bin_2_natural))/length(rr_14d_bin$rr_bin_2_natural),
                                       length(which(rr_14d_bin$rr_bin_3_natural))/length(rr_14d_bin$rr_bin_3_natural),
                                       length(which(rr_14d_bin$rr_bin_4_natural))/length(rr_14d_bin$rr_bin_4_natural),
                                       length(which(rr_14d_bin$rr_bin_5_natural))/length(rr_14d_bin$rr_bin_5_natural))  

rr_14d_prob_yr <-rbind(rr_14d_prob_yr, (cbind(length(which(rr_14d_bin$rr_bin_1_scenario1))/length(rr_14d_bin$rr_bin_1_scenario1),
                                       length(which(rr_14d_bin$rr_bin_2_scenario1))/length(rr_14d_bin$rr_bin_2_scenario1),
                                       length(which(rr_14d_bin$rr_bin_3_scenario1))/length(rr_14d_bin$rr_bin_3_scenario1),
                                       length(which(rr_14d_bin$rr_bin_4_scenario1))/length(rr_14d_bin$rr_bin_4_scenario1),
                                       length(which(rr_14d_bin$rr_bin_5_scenario1))/length(rr_14d_bin$rr_bin_5_scenario1))))  

rr_14d_prob_yr <-rbind(rr_14d_prob_yr, (cbind(length(which(rr_14d_bin$rr_bin_1_scenario2))/length(rr_14d_bin$rr_bin_1_scenario2),
                                       length(which(rr_14d_bin$rr_bin_2_scenario2))/length(rr_14d_bin$rr_bin_2_scenario2),
                                       length(which(rr_14d_bin$rr_bin_3_scenario2))/length(rr_14d_bin$rr_bin_3_scenario2),
                                       length(which(rr_14d_bin$rr_bin_4_scenario2))/length(rr_14d_bin$rr_bin_4_scenario2),
                                       length(which(rr_14d_bin$rr_bin_5_scenario2))/length(rr_14d_bin$rr_bin_5_scenario2)))) 

rr_14d_prob_yr <-rbind(rr_14d_prob_yr, (cbind(length(which(rr_14d_bin$rr_bin_1_scenario3))/length(rr_14d_bin$rr_bin_1_scenario3),
                                       length(which(rr_14d_bin$rr_bin_2_scenario3))/length(rr_14d_bin$rr_bin_2_scenario3),
                                       length(which(rr_14d_bin$rr_bin_3_scenario3))/length(rr_14d_bin$rr_bin_3_scenario3),
                                       length(which(rr_14d_bin$rr_bin_4_scenario3))/length(rr_14d_bin$rr_bin_4_scenario3),
                                       length(which(rr_14d_bin$rr_bin_5_scenario3))/length(rr_14d_bin$rr_bin_5_scenario3))))

colnames(rr_14d_prob_yr) <- c("recessionrate_bin_1","recessionrate_bin_2", "recessionrate_bin_3","recessionrate_bin_4","recessionrate_bin_5")
rownames(rr_14d_prob_yr) <- c("natural(cm/d)","Alt_E(cm/d)", "Alt_B(cm/d)", "Alt_A(cm/d)")
rr_14d_prob_yr <- rr_14d_prob_yr/rowSums(rr_14d_prob_yr)
```

Calculates the probability of 14-day recession rate ranges for the natural scenario and discrete time state (April-May, June-July, August-September).
```{r}
apr_may_bin <- rr_14d_bin[months(rr_14d_bin$date) %in% month.name[4:5],2:21]
jun_jul_bin <- rr_14d_bin[months(rr_14d_bin$date) %in% month.name[6:7],2:21]
aug_sep_bin <- rr_14d_bin[months(rr_14d_bin$date) %in% month.name[8:9],2:21]

rr_prob_natural_discrete <- cbind(length(which(apr_may_bin$rr_bin_1_natural))/length(apr_may_bin$rr_bin_1_natural),
                                       length(which(apr_may_bin$rr_bin_2_natural))/length(apr_may_bin$rr_bin_2_natural),
                                       length(which(apr_may_bin$rr_bin_3_natural))/length(apr_may_bin$rr_bin_3_natural),
                                       length(which(apr_may_bin$rr_bin_4_natural))/length(apr_may_bin$rr_bin_4_natural),
                                       length(which(apr_may_bin$rr_bin_5_natural))/length(apr_may_bin$rr_bin_5_natural))
                                
rr_prob_natural_discrete <- rbind(rr_prob_natural_discrete,
                                (cbind(length(which(jun_jul_bin$rr_bin_1_natural))/length(jun_jul_bin$rr_bin_1_natural),
                                       length(which(jun_jul_bin$rr_bin_2_natural))/length(jun_jul_bin$rr_bin_2_natural),
                                       length(which(jun_jul_bin$rr_bin_3_natural))/length(jun_jul_bin$rr_bin_3_natural),
                                       length(which(jun_jul_bin$rr_bin_4_natural))/length(jun_jul_bin$rr_bin_4_natural),
                                       length(which(jun_jul_bin$rr_bin_5_natural))/length(jun_jul_bin$rr_bin_5_natural))))

rr_prob_natural_discrete <- rbind(rr_prob_natural_discrete,
                                (cbind(length(which(aug_sep_bin$rr_bin_1_natural))/length(aug_sep_bin$rr_bin_1_natural),
                                       length(which(aug_sep_bin$rr_bin_2_natural))/length(aug_sep_bin$rr_bin_2_natural),
                                       length(which(aug_sep_bin$rr_bin_3_natural))/length(aug_sep_bin$rr_bin_3_natural),
                                       length(which(aug_sep_bin$rr_bin_4_natural))/length(aug_sep_bin$rr_bin_4_natural),
                                       length(which(aug_sep_bin$rr_bin_5_natural))/length(aug_sep_bin$rr_bin_5_natural))))


colnames(rr_prob_natural_discrete) <- c("recessionrate_bin_1","recessionrate_bin_2", "recessionrate_bin_3","recessionrate_bin_4","recessionrate_bin_5")
rownames(rr_prob_natural_discrete) <- c("apr_may","jun_jul","aug_sep")
rr_prob_natural_discrete <- rr_prob_natural_discrete/rowSums(rr_prob_natural_discrete)
```

Calculates the probability of 14-day recession rate ranges for the scenario1 scenario and discrete time state (April-May, June-July, August-September).
```{r}
apr_may_bin <- rr_14d_bin[months(rr_14d_bin$date) %in% month.name[4:5],2:21]
jun_jul_bin <- rr_14d_bin[months(rr_14d_bin$date) %in% month.name[6:7],2:21]
aug_sep_bin <- rr_14d_bin[months(rr_14d_bin$date) %in% month.name[8:9],2:21]

rr_prob_scenario1_discrete <- cbind(length(which(apr_may_bin$rr_bin_1_scenario1))/length(apr_may_bin$rr_bin_1_scenario1),
                                       length(which(apr_may_bin$rr_bin_2_scenario1))/length(apr_may_bin$rr_bin_2_scenario1),
                                       length(which(apr_may_bin$rr_bin_3_scenario1))/length(apr_may_bin$rr_bin_3_scenario1),
                                       length(which(apr_may_bin$rr_bin_4_scenario1))/length(apr_may_bin$rr_bin_4_scenario1),
                                       length(which(apr_may_bin$rr_bin_5_scenario1))/length(apr_may_bin$rr_bin_5_scenario1))
                                
rr_prob_scenario1_discrete <- rbind(rr_prob_scenario1_discrete,
                                (cbind(length(which(jun_jul_bin$rr_bin_1_scenario1))/length(jun_jul_bin$rr_bin_1_scenario1),
                                       length(which(jun_jul_bin$rr_bin_2_scenario1))/length(jun_jul_bin$rr_bin_2_scenario1),
                                       length(which(jun_jul_bin$rr_bin_3_scenario1))/length(jun_jul_bin$rr_bin_3_scenario1),
                                       length(which(jun_jul_bin$rr_bin_4_scenario1))/length(jun_jul_bin$rr_bin_4_scenario1),
                                       length(which(jun_jul_bin$rr_bin_5_scenario1))/length(jun_jul_bin$rr_bin_5_scenario1))))

rr_prob_scenario1_discrete <- rbind(rr_prob_scenario1_discrete,
                                (cbind(length(which(aug_sep_bin$rr_bin_1_scenario1))/length(aug_sep_bin$rr_bin_1_scenario1),
                                       length(which(aug_sep_bin$rr_bin_2_scenario1))/length(aug_sep_bin$rr_bin_2_scenario1),
                                       length(which(aug_sep_bin$rr_bin_3_scenario1))/length(aug_sep_bin$rr_bin_3_scenario1),
                                       length(which(aug_sep_bin$rr_bin_4_scenario1))/length(aug_sep_bin$rr_bin_4_scenario1),
                                       length(which(aug_sep_bin$rr_bin_5_scenario1))/length(aug_sep_bin$rr_bin_5_scenario1))))


colnames(rr_prob_scenario1_discrete) <- c("recessionrate_bin_1","recessionrate_bin_2", "recessionrate_bin_3","recessionrate_bin_4","recessionrate_bin_5")
rownames(rr_prob_scenario1_discrete) <- c("apr_may","jun_jul","aug_sep")
rr_prob_scenario1_discrete <- rr_prob_scenario1_discrete/rowSums(rr_prob_scenario1_discrete)
```

Calculates the probability of 14-day recession rate ranges for the scenario2 scenario and discrete time state (April-May, June-July, August-September).
```{r}
apr_may_bin <- rr_14d_bin[months(rr_14d_bin$date) %in% month.name[4:5],2:21]
jun_jul_bin <- rr_14d_bin[months(rr_14d_bin$date) %in% month.name[6:7],2:21]
aug_sep_bin <- rr_14d_bin[months(rr_14d_bin$date) %in% month.name[8:9],2:21]

rr_prob_scenario2_discrete <- cbind(length(which(apr_may_bin$rr_bin_1_scenario2))/length(apr_may_bin$rr_bin_1_scenario2),
                                       length(which(apr_may_bin$rr_bin_2_scenario2))/length(apr_may_bin$rr_bin_2_scenario2),
                                       length(which(apr_may_bin$rr_bin_3_scenario2))/length(apr_may_bin$rr_bin_3_scenario2),
                                       length(which(apr_may_bin$rr_bin_4_scenario2))/length(apr_may_bin$rr_bin_4_scenario2),
                                       length(which(apr_may_bin$rr_bin_5_scenario2))/length(apr_may_bin$rr_bin_5_scenario2))
                                
rr_prob_scenario2_discrete <- rbind(rr_prob_scenario2_discrete,
                                (cbind(length(which(jun_jul_bin$rr_bin_1_scenario2))/length(jun_jul_bin$rr_bin_1_scenario2),
                                       length(which(jun_jul_bin$rr_bin_2_scenario2))/length(jun_jul_bin$rr_bin_2_scenario2),
                                       length(which(jun_jul_bin$rr_bin_3_scenario2))/length(jun_jul_bin$rr_bin_3_scenario2),
                                       length(which(jun_jul_bin$rr_bin_4_scenario2))/length(jun_jul_bin$rr_bin_4_scenario2),
                                       length(which(jun_jul_bin$rr_bin_5_scenario2))/length(jun_jul_bin$rr_bin_5_scenario2))))

rr_prob_scenario2_discrete <- rbind(rr_prob_scenario2_discrete,
                                (cbind(length(which(aug_sep_bin$rr_bin_1_scenario2))/length(aug_sep_bin$rr_bin_1_scenario2),
                                       length(which(aug_sep_bin$rr_bin_2_scenario2))/length(aug_sep_bin$rr_bin_2_scenario2),
                                       length(which(aug_sep_bin$rr_bin_3_scenario2))/length(aug_sep_bin$rr_bin_3_scenario2),
                                       length(which(aug_sep_bin$rr_bin_4_scenario2))/length(aug_sep_bin$rr_bin_4_scenario2),
                                       length(which(aug_sep_bin$rr_bin_5_scenario2))/length(aug_sep_bin$rr_bin_5_scenario2))))


colnames(rr_prob_scenario2_discrete) <- c("recessionrate_bin_1","recessionrate_bin_2", "recessionrate_bin_3","recessionrate_bin_4","recessionrate_bin_5")
rownames(rr_prob_scenario2_discrete) <- c("apr_may","jun_jul","aug_sep")
rr_prob_scenario2_discrete <- rr_prob_scenario2_discrete/rowSums(rr_prob_scenario2_discrete)
```

Calculates the probability of 14-day recession rate ranges for the scenario3 scenario and discrete time state (April-May, June-July, August-September).
```{r}
apr_may_bin <- rr_14d_bin[months(rr_14d_bin$date) %in% month.name[4:5],2:21]
jun_jul_bin <- rr_14d_bin[months(rr_14d_bin$date) %in% month.name[6:7],2:21]
aug_sep_bin <- rr_14d_bin[months(rr_14d_bin$date) %in% month.name[8:9],2:21]

rr_prob_scenario3_discrete <- cbind(length(which(apr_may_bin$rr_bin_1_scenario3))/length(apr_may_bin$rr_bin_1_scenario3),
                                       length(which(apr_may_bin$rr_bin_2_scenario3))/length(apr_may_bin$rr_bin_2_scenario3),
                                       length(which(apr_may_bin$rr_bin_3_scenario3))/length(apr_may_bin$rr_bin_3_scenario3),
                                       length(which(apr_may_bin$rr_bin_4_scenario3))/length(apr_may_bin$rr_bin_4_scenario3),
                                       length(which(apr_may_bin$rr_bin_5_scenario3))/length(apr_may_bin$rr_bin_5_scenario3))
                                
rr_prob_scenario3_discrete <- rbind(rr_prob_scenario3_discrete,
                                (cbind(length(which(jun_jul_bin$rr_bin_1_scenario3))/length(jun_jul_bin$rr_bin_1_scenario3),
                                       length(which(jun_jul_bin$rr_bin_2_scenario3))/length(jun_jul_bin$rr_bin_2_scenario3),
                                       length(which(jun_jul_bin$rr_bin_3_scenario3))/length(jun_jul_bin$rr_bin_3_scenario3),
                                       length(which(jun_jul_bin$rr_bin_4_scenario3))/length(jun_jul_bin$rr_bin_4_scenario3),
                                       length(which(jun_jul_bin$rr_bin_5_scenario3))/length(jun_jul_bin$rr_bin_5_scenario3))))

rr_prob_scenario3_discrete <- rbind(rr_prob_scenario3_discrete,
                                (cbind(length(which(aug_sep_bin$rr_bin_1_scenario3))/length(aug_sep_bin$rr_bin_1_scenario3),
                                       length(which(aug_sep_bin$rr_bin_2_scenario3))/length(aug_sep_bin$rr_bin_2_scenario3),
                                       length(which(aug_sep_bin$rr_bin_3_scenario3))/length(aug_sep_bin$rr_bin_3_scenario3),
                                       length(which(aug_sep_bin$rr_bin_4_scenario3))/length(aug_sep_bin$rr_bin_4_scenario3),
                                       length(which(aug_sep_bin$rr_bin_5_scenario3))/length(aug_sep_bin$rr_bin_5_scenario3))))


colnames(rr_prob_scenario3_discrete) <- c("recessionrate_bin_1","recessionrate_bin_2", "recessionrate_bin_3","recessionrate_bin_4","recessionrate_bin_5")
rownames(rr_prob_scenario3_discrete) <- c("apr_may","jun_jul","aug_sep")
rr_prob_scenario3_discrete <- rr_prob_scenario3_discrete/rowSums(rr_prob_scenario3_discrete)
```

#### Process Q-model bins ####
Establishes model q-bins and calculates the probability of discharge range for any given entire year and discrete time state (April-May, June-July, August-September).Model Q-bins were origninal selected so that inundation areas increased by less than 15% between the discharge associated with each bin (i.e. 1000 cfs, 1500 cfs, 2000 cfs, 2500 cfs, 3000 cfs, 3500 cfs, 4000 cfs). The number of bins were then increased based on a frequency analysis adding 8 more bins for a total of 15 bins (i.e.200 cfs,225 cfs,250 cfs,275 cfs,300 cfs,350 cfs,400 cfs,500 cfs,1000 cfs, 1500 cfs, 2000 cfs, 2500 cfs, 3000 cfs, 3500 cfs, 4000 cfs).

Calculates the annual probability of a bin flood event occuring for all four flow scenarios(i.e.1-natural, 2-Alt. E, 3-Alt.B , and 4-Alt.A (no action)).
```{r}
q_bin_natural <-cbind.data.frame(hydro$date,
              q_bin_0 <- hydro[,2]< q_bin[1],
              q_bin_1 <- q_bin[1]<= hydro[,2] & hydro[,2]<q_bin[2],
              q_bin_2 <- q_bin[2]<= hydro[,2] & hydro[,2]<q_bin[3],
              q_bin_3 <- q_bin[3]<= hydro[,2] & hydro[,2]<q_bin[4],
              q_bin_4 <- q_bin[4]<= hydro[,2] & hydro[,2]<q_bin[5],
              q_bin_5 <- q_bin[5]<= hydro[,2] & hydro[,2]<q_bin[6],
              q_bin_6 <- q_bin[6]<= hydro[,2] & hydro[,2]<q_bin[7],
              q_bin_7 <- q_bin[7]<= hydro[,2] & hydro[,2]<q_bin[8],
              q_bin_8 <- q_bin[8]<= hydro[,2] & hydro[,2]<q_bin[9],
              q_bin_9 <- q_bin[9]<= hydro[,2] & hydro[,2]<q_bin[10],
              q_bin_10 <- q_bin[10]<= hydro[,2] & hydro[,2]<q_bin[11],
              q_bin_11 <- q_bin[11]<= hydro[,2] & hydro[,2]<q_bin[12],
              q_bin_12 <- q_bin[12]<= hydro[,2] & hydro[,2]<q_bin[13],
              q_bin_13 <- q_bin[13]<= hydro[,2] & hydro[,2]<q_bin[14],
              q_bin_14 <- q_bin[14]<= hydro[,2] & hydro[,2]<q_bin[15],
              q_bin_15 <- hydro[,2]>q_bin[15])

colnames(q_bin_natural) <- c("date","q_bin_0","q_bin_1", "q_bin_2","q_bin_3","q_bin_4","q_bin_5","q_bin_6","q_bin_7",
                             "q_bin_8", "q_bin_9","q_bin_10","q_bin_11","q_bin_12","q_bin_13","q_bin_14","q_bin_15")

q_prob_natural_yr <- cbind(length(which(q_bin_1))/(length(q_bin_1)-length(which(q_bin_0))),
                                       length(which(q_bin_2))/(length(q_bin_2)-length(which(q_bin_0))),
                                       length(which(q_bin_3))/(length(q_bin_3)-length(which(q_bin_0))),
                                       length(which(q_bin_4))/(length(q_bin_4)-length(which(q_bin_0))),
                                       length(which(q_bin_5))/(length(q_bin_5)-length(which(q_bin_0))),
                                       length(which(q_bin_6))/(length(q_bin_6)-length(which(q_bin_0))),
                                       length(which(q_bin_7))/(length(q_bin_7)-length(which(q_bin_0))),
                                       length(which(q_bin_8))/(length(q_bin_8)-length(which(q_bin_0))),
                                       length(which(q_bin_9))/(length(q_bin_9)-length(which(q_bin_0))),
                                       length(which(q_bin_10))/(length(q_bin_10)-length(which(q_bin_0))),
                                       length(which(q_bin_11))/(length(q_bin_11)-length(which(q_bin_0))),
                                       length(which(q_bin_12))/(length(q_bin_12)-length(which(q_bin_0))),
                                       length(which(q_bin_13))/(length(q_bin_13)-length(which(q_bin_0))),
                                       length(which(q_bin_14))/(length(q_bin_14)-length(which(q_bin_0))),
                                       length(which(q_bin_15))/(length(q_bin_15)-length(which(q_bin_0))))

colnames(q_prob_natural_yr) <- c("q_bin_1", "q_bin_2","q_bin_3","q_bin_4","q_bin_5","q_bin_6","q_bin_7",
                             "q_bin_8", "q_bin_9","q_bin_10","q_bin_11","q_bin_12","q_bin_13","q_bin_14","q_bin_15")
row.names(q_prob_natural_yr) <- c("q_prob")

q_bin_scenario1 <-cbind.data.frame(hydro$date,
              q_bin_0 <- hydro[,3]< q_bin[1],
              q_bin_1 <- q_bin[1]<= hydro[,3] & hydro[,3]<q_bin[2],
              q_bin_2 <- q_bin[2]<= hydro[,3] & hydro[,3]<q_bin[3],
              q_bin_3 <- q_bin[3]<= hydro[,3] & hydro[,3]<q_bin[4],
              q_bin_4 <- q_bin[4]<= hydro[,3] & hydro[,3]<q_bin[5],
              q_bin_5 <- q_bin[5]<= hydro[,3] & hydro[,3]<q_bin[6],
              q_bin_6 <- q_bin[6]<= hydro[,3] & hydro[,3]<q_bin[7],
              q_bin_7 <- q_bin[7]<= hydro[,3] & hydro[,3]<q_bin[8],
              q_bin_8 <- q_bin[8]<= hydro[,3] & hydro[,3]<q_bin[9],
              q_bin_9 <- q_bin[9]<= hydro[,3] & hydro[,3]<q_bin[10],
              q_bin_10 <- q_bin[10]<= hydro[,3] & hydro[,3]<q_bin[11],
              q_bin_11 <- q_bin[11]<= hydro[,3] & hydro[,3]<q_bin[12],
              q_bin_12 <- q_bin[12]<= hydro[,3] & hydro[,3]<q_bin[13],
              q_bin_13 <- q_bin[13]<= hydro[,3] & hydro[,3]<q_bin[14],
              q_bin_14 <- q_bin[14]<= hydro[,3] & hydro[,3]<q_bin[15],
              q_bin_15 <- hydro[,3]>q_bin[15])

colnames(q_bin_scenario1) <- c("date","q_bin_0","q_bin_1", "q_bin_2","q_bin_3","q_bin_4","q_bin_5","q_bin_6","q_bin_7",
                             "q_bin_8", "q_bin_9","q_bin_10","q_bin_11","q_bin_12","q_bin_13","q_bin_14","q_bin_15")

q_prob_scenario1_yr <- cbind(length(which(q_bin_1))/(length(q_bin_1)-length(which(q_bin_0))),
                                       length(which(q_bin_2))/(length(q_bin_2)-length(which(q_bin_0))),
                                       length(which(q_bin_3))/(length(q_bin_3)-length(which(q_bin_0))),
                                       length(which(q_bin_4))/(length(q_bin_4)-length(which(q_bin_0))),
                                       length(which(q_bin_5))/(length(q_bin_5)-length(which(q_bin_0))),
                                       length(which(q_bin_6))/(length(q_bin_6)-length(which(q_bin_0))),
                                       length(which(q_bin_7))/(length(q_bin_7)-length(which(q_bin_0))),
                                       length(which(q_bin_8))/(length(q_bin_8)-length(which(q_bin_0))),
                                       length(which(q_bin_9))/(length(q_bin_9)-length(which(q_bin_0))),
                                       length(which(q_bin_10))/(length(q_bin_10)-length(which(q_bin_0))),
                                       length(which(q_bin_11))/(length(q_bin_11)-length(which(q_bin_0))),
                                       length(which(q_bin_12))/(length(q_bin_12)-length(which(q_bin_0))),
                                       length(which(q_bin_13))/(length(q_bin_13)-length(which(q_bin_0))),
                                       length(which(q_bin_14))/(length(q_bin_14)-length(which(q_bin_0))),
                                       length(which(q_bin_15))/(length(q_bin_15)-length(which(q_bin_0))))

colnames(q_prob_scenario1_yr) <- c("q_bin_1", "q_bin_2","q_bin_3","q_bin_4","q_bin_5","q_bin_6","q_bin_7",
                                   "q_bin_8", "q_bin_9","q_bin_10","q_bin_11","q_bin_12","q_bin_13","q_bin_14","q_bin_15")
row.names(q_prob_scenario1_yr) <- c("q_prob")

q_bin_scenario2 <-cbind.data.frame(hydro$date,
              q_bin_0 <- hydro[,4]< q_bin[1],
              q_bin_1 <- q_bin[1]<= hydro[,4] & hydro[,4]<q_bin[2],
              q_bin_2 <- q_bin[2]<= hydro[,4] & hydro[,4]<q_bin[3],
              q_bin_3 <- q_bin[3]<= hydro[,4] & hydro[,4]<q_bin[4],
              q_bin_4 <- q_bin[4]<= hydro[,4] & hydro[,4]<q_bin[5],
              q_bin_5 <- q_bin[5]<= hydro[,4] & hydro[,4]<q_bin[6],
              q_bin_6 <- q_bin[6]<= hydro[,4] & hydro[,4]<q_bin[7],
              q_bin_7 <- q_bin[7]<= hydro[,4] & hydro[,4]<q_bin[8],
              q_bin_8 <- q_bin[8]<= hydro[,4] & hydro[,4]<q_bin[9],
              q_bin_9 <- q_bin[9]<= hydro[,4] & hydro[,4]<q_bin[10],
              q_bin_10 <- q_bin[10]<= hydro[,4] & hydro[,4]<q_bin[11],
              q_bin_11 <- q_bin[11]<= hydro[,4] & hydro[,4]<q_bin[12],
              q_bin_12 <- q_bin[12]<= hydro[,4] & hydro[,4]<q_bin[13],
              q_bin_13 <- q_bin[13]<= hydro[,4] & hydro[,4]<q_bin[14],
              q_bin_14 <- q_bin[14]<= hydro[,4] & hydro[,4]<q_bin[15],
              q_bin_15 <- hydro[,4]>q_bin[15])

colnames(q_bin_scenario2) <- c("date","q_bin_0","q_bin_1", "q_bin_2","q_bin_3","q_bin_4","q_bin_5","q_bin_6","q_bin_7",
                               "q_bin_8", "q_bin_9","q_bin_10","q_bin_11","q_bin_12","q_bin_13","q_bin_14","q_bin_15")

q_prob_scenario2_yr <- cbind(length(which(q_bin_1))/(length(q_bin_1)-length(which(q_bin_0))),
                                       length(which(q_bin_2))/(length(q_bin_2)-length(which(q_bin_0))),
                                       length(which(q_bin_3))/(length(q_bin_3)-length(which(q_bin_0))),
                                       length(which(q_bin_4))/(length(q_bin_4)-length(which(q_bin_0))),
                                       length(which(q_bin_5))/(length(q_bin_5)-length(which(q_bin_0))),
                                       length(which(q_bin_6))/(length(q_bin_6)-length(which(q_bin_0))),
                                       length(which(q_bin_7))/(length(q_bin_7)-length(which(q_bin_0))),
                                       length(which(q_bin_8))/(length(q_bin_8)-length(which(q_bin_0))),
                                       length(which(q_bin_9))/(length(q_bin_9)-length(which(q_bin_0))),
                                       length(which(q_bin_10))/(length(q_bin_10)-length(which(q_bin_0))),
                                       length(which(q_bin_11))/(length(q_bin_11)-length(which(q_bin_0))),
                                       length(which(q_bin_12))/(length(q_bin_12)-length(which(q_bin_0))),
                                       length(which(q_bin_13))/(length(q_bin_13)-length(which(q_bin_0))),
                                       length(which(q_bin_14))/(length(q_bin_14)-length(which(q_bin_0))),
                                       length(which(q_bin_15))/(length(q_bin_15)-length(which(q_bin_0))))

colnames(q_prob_scenario2_yr) <- c("q_bin_1", "q_bin_2","q_bin_3","q_bin_4","q_bin_5","q_bin_6","q_bin_7",
                                   "q_bin_8", "q_bin_9","q_bin_10","q_bin_11","q_bin_12","q_bin_13","q_bin_14","q_bin_15")
row.names(q_prob_scenario2_yr) <- c("q_prob")

q_bin_scenario3 <-cbind.data.frame(hydro$date,
              q_bin_0 <- hydro[,5]< q_bin[1],
              q_bin_1 <- q_bin[1]<= hydro[,5] & hydro[,5]<q_bin[2],
              q_bin_2 <- q_bin[2]<= hydro[,5] & hydro[,5]<q_bin[3],
              q_bin_3 <- q_bin[3]<= hydro[,5] & hydro[,5]<q_bin[4],
              q_bin_4 <- q_bin[4]<= hydro[,5] & hydro[,5]<q_bin[5],
              q_bin_5 <- q_bin[5]<= hydro[,5] & hydro[,5]<q_bin[6],
              q_bin_6 <- q_bin[6]<= hydro[,5] & hydro[,5]<q_bin[7],
              q_bin_7 <- q_bin[7]<= hydro[,5] & hydro[,5]<q_bin[8],
              q_bin_8 <- q_bin[8]<= hydro[,5] & hydro[,5]<q_bin[9],
              q_bin_9 <- q_bin[9]<= hydro[,5] & hydro[,5]<q_bin[10],
              q_bin_10 <- q_bin[10]<= hydro[,5] & hydro[,5]<q_bin[11],
              q_bin_11 <- q_bin[11]<= hydro[,5] & hydro[,5]<q_bin[12],
              q_bin_12 <- q_bin[12]<= hydro[,5] & hydro[,5]<q_bin[13],
              q_bin_13 <- q_bin[13]<= hydro[,5] & hydro[,5]<q_bin[14],
              q_bin_14 <- q_bin[14]<= hydro[,5] & hydro[,5]<q_bin[15],
              q_bin_15 <- hydro[,5]>q_bin[15])

colnames(q_bin_scenario3) <- c("date","q_bin_0","q_bin_1", "q_bin_2","q_bin_3","q_bin_4","q_bin_5","q_bin_6","q_bin_7",
                             "q_bin_8", "q_bin_9","q_bin_10","q_bin_11","q_bin_12","q_bin_13","q_bin_14","q_bin_15")

q_prob_scenario3_yr <- cbind(length(which(q_bin_1))/(length(q_bin_1)-length(which(q_bin_0))),
                                       length(which(q_bin_2))/(length(q_bin_2)-length(which(q_bin_0))),
                                       length(which(q_bin_3))/(length(q_bin_3)-length(which(q_bin_0))),
                                       length(which(q_bin_4))/(length(q_bin_4)-length(which(q_bin_0))),
                                       length(which(q_bin_5))/(length(q_bin_5)-length(which(q_bin_0))),
                                       length(which(q_bin_6))/(length(q_bin_6)-length(which(q_bin_0))),
                                       length(which(q_bin_7))/(length(q_bin_7)-length(which(q_bin_0))),
                                       length(which(q_bin_8))/(length(q_bin_8)-length(which(q_bin_0))),
                                       length(which(q_bin_9))/(length(q_bin_9)-length(which(q_bin_0))),
                                       length(which(q_bin_10))/(length(q_bin_10)-length(which(q_bin_0))),
                                       length(which(q_bin_11))/(length(q_bin_11)-length(which(q_bin_0))),
                                       length(which(q_bin_12))/(length(q_bin_12)-length(which(q_bin_0))),
                                       length(which(q_bin_13))/(length(q_bin_13)-length(which(q_bin_0))),
                                       length(which(q_bin_14))/(length(q_bin_14)-length(which(q_bin_0))),
                                       length(which(q_bin_15))/(length(q_bin_15)-length(which(q_bin_0))))

colnames(q_prob_scenario3_yr) <- c("q_bin_1", "q_bin_2","q_bin_3","q_bin_4","q_bin_5","q_bin_6","q_bin_7",
                                    "q_bin_8", "q_bin_9","q_bin_10","q_bin_11","q_bin_12","q_bin_13","q_bin_14","q_bin_15")
row.names(q_prob_scenario3_yr) <- c("q_prob")

q_prob_yr <- rbind(q_prob_natural_yr,q_prob_scenario2_yr,q_prob_scenario2_yr,q_prob_scenario3_yr)
row.names(q_prob_yr) <- c("natural", "Alt_E", "Alt_B", "scenario3")
q_prob_yr <- q_prob_yr/rowSums(q_prob_yr)
```

Calculates the probability of a bin flood event occuring for the natural flow scenario for discrete time states (April-May, June-July, August-September). Each flow scenario corrects the probability distribution for error by normalizing the probabilities (i.e. ensuring the distribution sums to 1).
```{r}
apr_may_bin <- q_bin_natural[months(q_bin_natural$date) %in% month.name[4:5],]
jun_jul_bin <- q_bin_natural[months(q_bin_natural$date) %in% month.name[6:7],]
aug_sep_bin <- q_bin_natural[months(q_bin_natural$date) %in% month.name[8:9],]

q_prob_natural_apr_may <- cbind(length(which(apr_may_bin$q_bin_1))/(length(apr_may_bin$q_bin_1)-length(which(apr_may_bin$q_bin_0))),
                                       length(which(apr_may_bin$q_bin_2))/(length(apr_may_bin$q_bin_2)-length(which(apr_may_bin$q_bin_0))),
                                       length(which(apr_may_bin$q_bin_3))/(length(apr_may_bin$q_bin_3)-length(which(apr_may_bin$q_bin_0))),
                                       length(which(apr_may_bin$q_bin_4))/(length(apr_may_bin$q_bin_4)-length(which(apr_may_bin$q_bin_0))),
                                       length(which(apr_may_bin$q_bin_5))/(length(apr_may_bin$q_bin_5)-length(which(apr_may_bin$q_bin_0))),
                                       length(which(apr_may_bin$q_bin_6))/(length(apr_may_bin$q_bin_6)-length(which(apr_may_bin$q_bin_0))),
                                       length(which(apr_may_bin$q_bin_7))/(length(apr_may_bin$q_bin_7)-length(which(apr_may_bin$q_bin_0))),
                                       length(which(apr_may_bin$q_bin_8))/(length(apr_may_bin$q_bin_8)-length(which(apr_may_bin$q_bin_0))),
                                       length(which(apr_may_bin$q_bin_9))/(length(apr_may_bin$q_bin_9)-length(which(apr_may_bin$q_bin_0))),
                                       length(which(apr_may_bin$q_bin_10))/(length(apr_may_bin$q_bin_10)-length(which(apr_may_bin$q_bin_0))),
                                       length(which(apr_may_bin$q_bin_11))/(length(apr_may_bin$q_bin_11)-length(which(apr_may_bin$q_bin_0))),
                                       length(which(apr_may_bin$q_bin_12))/(length(apr_may_bin$q_bin_12)-length(which(apr_may_bin$q_bin_0))),
                                       length(which(apr_may_bin$q_bin_13))/(length(apr_may_bin$q_bin_13)-length(which(apr_may_bin$q_bin_0))),
                                       length(which(apr_may_bin$q_bin_14))/(length(apr_may_bin$q_bin_14)-length(which(apr_may_bin$q_bin_0))),
                                       length(which(apr_may_bin$q_bin_15))/(length(apr_may_bin$q_bin_15)-length(which(apr_may_bin$q_bin_0))))

colnames(q_prob_natural_apr_may) <- c("q_bin_1", "q_bin_2","q_bin_3","q_bin_4","q_bin_5","q_bin_6","q_bin_7",
                                    "q_bin_8", "q_bin_9","q_bin_10","q_bin_11","q_bin_12","q_bin_13","q_bin_14","q_bin_15")

q_prob_natural_jun_jul <- cbind(length(which(jun_jul_bin$q_bin_1))/(length(jun_jul_bin$q_bin_1)-length(which(jun_jul_bin$q_bin_0))),
                                       length(which(jun_jul_bin$q_bin_2))/(length(jun_jul_bin$q_bin_2)-length(which(jun_jul_bin$q_bin_0))),
                                       length(which(jun_jul_bin$q_bin_3))/(length(jun_jul_bin$q_bin_3)-length(which(jun_jul_bin$q_bin_0))),
                                       length(which(jun_jul_bin$q_bin_4))/(length(jun_jul_bin$q_bin_4)-length(which(jun_jul_bin$q_bin_0))),
                                       length(which(jun_jul_bin$q_bin_5))/(length(jun_jul_bin$q_bin_5)-length(which(jun_jul_bin$q_bin_0))),
                                       length(which(jun_jul_bin$q_bin_6))/(length(jun_jul_bin$q_bin_6)-length(which(jun_jul_bin$q_bin_0))),
                                       length(which(jun_jul_bin$q_bin_7))/(length(jun_jul_bin$q_bin_7)-length(which(jun_jul_bin$q_bin_0))),
                                       length(which(jun_jul_bin$q_bin_8))/(length(jun_jul_bin$q_bin_8)-length(which(jun_jul_bin$q_bin_0))),
                                       length(which(jun_jul_bin$q_bin_9))/(length(jun_jul_bin$q_bin_9)-length(which(jun_jul_bin$q_bin_0))),
                                       length(which(jun_jul_bin$q_bin_10))/(length(jun_jul_bin$q_bin_10)-length(which(jun_jul_bin$q_bin_0))),
                                       length(which(jun_jul_bin$q_bin_11))/(length(jun_jul_bin$q_bin_11)-length(which(jun_jul_bin$q_bin_0))),
                                       length(which(jun_jul_bin$q_bin_12))/(length(jun_jul_bin$q_bin_12)-length(which(jun_jul_bin$q_bin_0))),
                                       length(which(jun_jul_bin$q_bin_13))/(length(jun_jul_bin$q_bin_13)-length(which(jun_jul_bin$q_bin_0))),
                                       length(which(jun_jul_bin$q_bin_14))/(length(jun_jul_bin$q_bin_14)-length(which(jun_jul_bin$q_bin_0))),
                                       length(which(jun_jul_bin$q_bin_15))/(length(jun_jul_bin$q_bin_15)-length(which(jun_jul_bin$q_bin_0))))

colnames(q_prob_natural_jun_jul) <- c("q_bin_1", "q_bin_2","q_bin_3","q_bin_4","q_bin_5","q_bin_6","q_bin_7",
                                    "q_bin_8", "q_bin_9","q_bin_10","q_bin_11","q_bin_12","q_bin_13","q_bin_14","q_bin_15")

q_prob_natural_aug_sep <- cbind(length(which(aug_sep_bin$q_bin_1))/(length(aug_sep_bin$q_bin_1)-length(which(aug_sep_bin$q_bin_0))),
                                       length(which(aug_sep_bin$q_bin_2))/(length(aug_sep_bin$q_bin_2)-length(which(aug_sep_bin$q_bin_0))),
                                       length(which(aug_sep_bin$q_bin_3))/(length(aug_sep_bin$q_bin_3)-length(which(aug_sep_bin$q_bin_0))),
                                       length(which(aug_sep_bin$q_bin_4))/(length(aug_sep_bin$q_bin_4)-length(which(aug_sep_bin$q_bin_0))),
                                       length(which(aug_sep_bin$q_bin_5))/(length(aug_sep_bin$q_bin_5)-length(which(aug_sep_bin$q_bin_0))),
                                       length(which(aug_sep_bin$q_bin_6))/(length(aug_sep_bin$q_bin_6)-length(which(aug_sep_bin$q_bin_0))),
                                       length(which(aug_sep_bin$q_bin_7))/(length(aug_sep_bin$q_bin_7)-length(which(aug_sep_bin$q_bin_0))),
                                       length(which(aug_sep_bin$q_bin_8))/(length(aug_sep_bin$q_bin_8)-length(which(aug_sep_bin$q_bin_0))),
                                       length(which(aug_sep_bin$q_bin_9))/(length(aug_sep_bin$q_bin_9)-length(which(aug_sep_bin$q_bin_0))),
                                       length(which(aug_sep_bin$q_bin_10))/(length(aug_sep_bin$q_bin_10)-length(which(aug_sep_bin$q_bin_0))),
                                       length(which(aug_sep_bin$q_bin_11))/(length(aug_sep_bin$q_bin_11)-length(which(aug_sep_bin$q_bin_0))),
                                       length(which(aug_sep_bin$q_bin_12))/(length(aug_sep_bin$q_bin_12)-length(which(aug_sep_bin$q_bin_0))),
                                       length(which(aug_sep_bin$q_bin_13))/(length(aug_sep_bin$q_bin_13)-length(which(aug_sep_bin$q_bin_0))),
                                       length(which(aug_sep_bin$q_bin_14))/(length(aug_sep_bin$q_bin_14)-length(which(aug_sep_bin$q_bin_0))),
                                       length(which(aug_sep_bin$q_bin_15))/(length(aug_sep_bin$q_bin_15)-length(which(aug_sep_bin$q_bin_0))))

colnames(q_prob_natural_aug_sep) <- c("q_bin_1", "q_bin_2","q_bin_3","q_bin_4","q_bin_5","q_bin_6","q_bin_7",
                                    "q_bin_8", "q_bin_9","q_bin_10","q_bin_11","q_bin_12","q_bin_13","q_bin_14","q_bin_15")

q_prob_natural_discrete <- rbind(q_prob_natural_apr_may,q_prob_natural_jun_jul,q_prob_natural_aug_sep)
row.names(q_prob_natural_discrete) <- c("apr_may","jun_jul","aug_sep")
q_prob_natural_discrete <- q_prob_natural_discrete/rowSums(q_prob_natural_discrete)
```

Calculates the probability of a bin flood event occuring for scenario2 flow scenario for discrete time states (April-May, June-July, August-September). Each flow scenario corrects the probability distribution for error by normalizing the probabilities (i.e. ensuring the distribution sums to 1).
```{r}
apr_may_bin <- q_bin_scenario1[months(q_bin_scenario1$date) %in% month.name[4:5],]
jun_jul_bin <- q_bin_scenario1[months(q_bin_scenario1$date) %in% month.name[6:7],]
aug_sep_bin <- q_bin_scenario1[months(q_bin_scenario1$date) %in% month.name[8:9],]

q_prob_scenario1_apr_may <- cbind(length(which(apr_may_bin$q_bin_1))/(length(apr_may_bin$q_bin_1)-length(which(apr_may_bin$q_bin_0))),
                                       length(which(apr_may_bin$q_bin_2))/(length(apr_may_bin$q_bin_2)-length(which(apr_may_bin$q_bin_0))),
                                       length(which(apr_may_bin$q_bin_3))/(length(apr_may_bin$q_bin_3)-length(which(apr_may_bin$q_bin_0))),
                                       length(which(apr_may_bin$q_bin_4))/(length(apr_may_bin$q_bin_4)-length(which(apr_may_bin$q_bin_0))),
                                       length(which(apr_may_bin$q_bin_5))/(length(apr_may_bin$q_bin_5)-length(which(apr_may_bin$q_bin_0))),
                                       length(which(apr_may_bin$q_bin_6))/(length(apr_may_bin$q_bin_6)-length(which(apr_may_bin$q_bin_0))),
                                       length(which(apr_may_bin$q_bin_7))/(length(apr_may_bin$q_bin_7)-length(which(apr_may_bin$q_bin_0))),
                                       length(which(apr_may_bin$q_bin_8))/(length(apr_may_bin$q_bin_8)-length(which(apr_may_bin$q_bin_0))),
                                       length(which(apr_may_bin$q_bin_9))/(length(apr_may_bin$q_bin_9)-length(which(apr_may_bin$q_bin_0))),
                                       length(which(apr_may_bin$q_bin_10))/(length(apr_may_bin$q_bin_10)-length(which(apr_may_bin$q_bin_0))),
                                       length(which(apr_may_bin$q_bin_11))/(length(apr_may_bin$q_bin_11)-length(which(apr_may_bin$q_bin_0))),
                                       length(which(apr_may_bin$q_bin_12))/(length(apr_may_bin$q_bin_12)-length(which(apr_may_bin$q_bin_0))),
                                       length(which(apr_may_bin$q_bin_13))/(length(apr_may_bin$q_bin_13)-length(which(apr_may_bin$q_bin_0))),
                                       length(which(apr_may_bin$q_bin_14))/(length(apr_may_bin$q_bin_14)-length(which(apr_may_bin$q_bin_0))),
                                       length(which(apr_may_bin$q_bin_15))/(length(apr_may_bin$q_bin_15)-length(which(apr_may_bin$q_bin_0))))

colnames(q_prob_scenario1_apr_may) <- c("q_bin_1", "q_bin_2","q_bin_3","q_bin_4","q_bin_5","q_bin_6","q_bin_7",
                                    "q_bin_8", "q_bin_9","q_bin_10","q_bin_11","q_bin_12","q_bin_13","q_bin_14","q_bin_15")

q_prob_scenario1_jun_jul <- cbind(length(which(jun_jul_bin$q_bin_1))/(length(jun_jul_bin$q_bin_1)-length(which(jun_jul_bin$q_bin_0))),
                                       length(which(jun_jul_bin$q_bin_2))/(length(jun_jul_bin$q_bin_2)-length(which(jun_jul_bin$q_bin_0))),
                                       length(which(jun_jul_bin$q_bin_3))/(length(jun_jul_bin$q_bin_3)-length(which(jun_jul_bin$q_bin_0))),
                                       length(which(jun_jul_bin$q_bin_4))/(length(jun_jul_bin$q_bin_4)-length(which(jun_jul_bin$q_bin_0))),
                                       length(which(jun_jul_bin$q_bin_5))/(length(jun_jul_bin$q_bin_5)-length(which(jun_jul_bin$q_bin_0))),
                                       length(which(jun_jul_bin$q_bin_6))/(length(jun_jul_bin$q_bin_6)-length(which(jun_jul_bin$q_bin_0))),
                                       length(which(jun_jul_bin$q_bin_7))/(length(jun_jul_bin$q_bin_7)-length(which(jun_jul_bin$q_bin_0))),
                                       length(which(jun_jul_bin$q_bin_8))/(length(jun_jul_bin$q_bin_8)-length(which(jun_jul_bin$q_bin_0))),
                                       length(which(jun_jul_bin$q_bin_9))/(length(jun_jul_bin$q_bin_9)-length(which(jun_jul_bin$q_bin_0))),
                                       length(which(jun_jul_bin$q_bin_10))/(length(jun_jul_bin$q_bin_10)-length(which(jun_jul_bin$q_bin_0))),
                                       length(which(jun_jul_bin$q_bin_11))/(length(jun_jul_bin$q_bin_11)-length(which(jun_jul_bin$q_bin_0))),
                                       length(which(jun_jul_bin$q_bin_12))/(length(jun_jul_bin$q_bin_12)-length(which(jun_jul_bin$q_bin_0))),
                                       length(which(jun_jul_bin$q_bin_13))/(length(jun_jul_bin$q_bin_13)-length(which(jun_jul_bin$q_bin_0))),
                                       length(which(jun_jul_bin$q_bin_14))/(length(jun_jul_bin$q_bin_14)-length(which(jun_jul_bin$q_bin_0))),
                                       length(which(jun_jul_bin$q_bin_15))/(length(jun_jul_bin$q_bin_15)-length(which(jun_jul_bin$q_bin_0))))

colnames(q_prob_scenario1_jun_jul) <- c("q_bin_1", "q_bin_2","q_bin_3","q_bin_4","q_bin_5","q_bin_6","q_bin_7",
                                    "q_bin_8", "q_bin_9","q_bin_10","q_bin_11","q_bin_12","q_bin_13","q_bin_14","q_bin_15")

q_prob_scenario1_aug_sep <- cbind(length(which(aug_sep_bin$q_bin_1))/(length(aug_sep_bin$q_bin_1)-length(which(aug_sep_bin$q_bin_0))),
                                       length(which(aug_sep_bin$q_bin_2))/(length(aug_sep_bin$q_bin_2)-length(which(aug_sep_bin$q_bin_0))),
                                       length(which(aug_sep_bin$q_bin_3))/(length(aug_sep_bin$q_bin_3)-length(which(aug_sep_bin$q_bin_0))),
                                       length(which(aug_sep_bin$q_bin_4))/(length(aug_sep_bin$q_bin_4)-length(which(aug_sep_bin$q_bin_0))),
                                       length(which(aug_sep_bin$q_bin_5))/(length(aug_sep_bin$q_bin_5)-length(which(aug_sep_bin$q_bin_0))),
                                       length(which(aug_sep_bin$q_bin_6))/(length(aug_sep_bin$q_bin_6)-length(which(aug_sep_bin$q_bin_0))),
                                       length(which(aug_sep_bin$q_bin_7))/(length(aug_sep_bin$q_bin_7)-length(which(aug_sep_bin$q_bin_0))),
                                       length(which(aug_sep_bin$q_bin_8))/(length(aug_sep_bin$q_bin_8)-length(which(aug_sep_bin$q_bin_0))),
                                       length(which(aug_sep_bin$q_bin_9))/(length(aug_sep_bin$q_bin_9)-length(which(aug_sep_bin$q_bin_0))),
                                       length(which(aug_sep_bin$q_bin_10))/(length(aug_sep_bin$q_bin_10)-length(which(aug_sep_bin$q_bin_0))),
                                       length(which(aug_sep_bin$q_bin_11))/(length(aug_sep_bin$q_bin_11)-length(which(aug_sep_bin$q_bin_0))),
                                       length(which(aug_sep_bin$q_bin_12))/(length(aug_sep_bin$q_bin_12)-length(which(aug_sep_bin$q_bin_0))),
                                       length(which(aug_sep_bin$q_bin_13))/(length(aug_sep_bin$q_bin_13)-length(which(aug_sep_bin$q_bin_0))),
                                       length(which(aug_sep_bin$q_bin_14))/(length(aug_sep_bin$q_bin_14)-length(which(aug_sep_bin$q_bin_0))),
                                       length(which(aug_sep_bin$q_bin_15))/(length(aug_sep_bin$q_bin_15)-length(which(aug_sep_bin$q_bin_0))))

colnames(q_prob_scenario1_aug_sep) <- c("q_bin_1", "q_bin_2","q_bin_3","q_bin_4","q_bin_5","q_bin_6","q_bin_7",
                                    "q_bin_8", "q_bin_9","q_bin_10","q_bin_11","q_bin_12","q_bin_13","q_bin_14","q_bin_15")

q_prob_scenario1_discrete <- rbind(q_prob_scenario1_apr_may,q_prob_scenario1_jun_jul,q_prob_scenario1_aug_sep)
row.names(q_prob_scenario1_discrete) <- c("apr_may","jun_jul","aug_sep")
q_prob_scenario1_discrete <- q_prob_scenario1_discrete/rowSums(q_prob_scenario1_discrete)
```

Calculates the probability of a bin flood event occuring for scenario2 flow scenario for discrete time states (April-May, June-July, August-September). Each flow scenario corrects the probability distribution for error by normalizing the probabilities (i.e. ensuring the distribution sums to 1).
```{r}
apr_may_bin <- q_bin_scenario2[months(q_bin_scenario2$date) %in% month.name[4:5],]
jun_jul_bin <- q_bin_scenario2[months(q_bin_scenario2$date) %in% month.name[6:7],]
aug_sep_bin <- q_bin_scenario2[months(q_bin_scenario2$date) %in% month.name[8:9],]

q_prob_scenario2_apr_may <- cbind(length(which(apr_may_bin$q_bin_1))/(length(apr_may_bin$q_bin_1)-length(which(apr_may_bin$q_bin_0))),
                                       length(which(apr_may_bin$q_bin_2))/(length(apr_may_bin$q_bin_2)-length(which(apr_may_bin$q_bin_0))),
                                       length(which(apr_may_bin$q_bin_3))/(length(apr_may_bin$q_bin_3)-length(which(apr_may_bin$q_bin_0))),
                                       length(which(apr_may_bin$q_bin_4))/(length(apr_may_bin$q_bin_4)-length(which(apr_may_bin$q_bin_0))),
                                       length(which(apr_may_bin$q_bin_5))/(length(apr_may_bin$q_bin_5)-length(which(apr_may_bin$q_bin_0))),
                                       length(which(apr_may_bin$q_bin_6))/(length(apr_may_bin$q_bin_6)-length(which(apr_may_bin$q_bin_0))),
                                       length(which(apr_may_bin$q_bin_7))/(length(apr_may_bin$q_bin_7)-length(which(apr_may_bin$q_bin_0))),
                                       length(which(apr_may_bin$q_bin_8))/(length(apr_may_bin$q_bin_8)-length(which(apr_may_bin$q_bin_0))),
                                       length(which(apr_may_bin$q_bin_9))/(length(apr_may_bin$q_bin_9)-length(which(apr_may_bin$q_bin_0))),
                                       length(which(apr_may_bin$q_bin_10))/(length(apr_may_bin$q_bin_10)-length(which(apr_may_bin$q_bin_0))),
                                       length(which(apr_may_bin$q_bin_11))/(length(apr_may_bin$q_bin_11)-length(which(apr_may_bin$q_bin_0))),
                                       length(which(apr_may_bin$q_bin_12))/(length(apr_may_bin$q_bin_12)-length(which(apr_may_bin$q_bin_0))),
                                       length(which(apr_may_bin$q_bin_13))/(length(apr_may_bin$q_bin_13)-length(which(apr_may_bin$q_bin_0))),
                                       length(which(apr_may_bin$q_bin_14))/(length(apr_may_bin$q_bin_14)-length(which(apr_may_bin$q_bin_0))),
                                       length(which(apr_may_bin$q_bin_15))/(length(apr_may_bin$q_bin_15)-length(which(apr_may_bin$q_bin_0))))

colnames(q_prob_scenario2_apr_may) <- c("q_bin_1", "q_bin_2","q_bin_3","q_bin_4","q_bin_5","q_bin_6","q_bin_7",
                                    "q_bin_8", "q_bin_9","q_bin_10","q_bin_11","q_bin_12","q_bin_13","q_bin_14","q_bin_15")

q_prob_scenario2_jun_jul <- cbind(length(which(jun_jul_bin$q_bin_1))/(length(jun_jul_bin$q_bin_1)-length(which(jun_jul_bin$q_bin_0))),
                                       length(which(jun_jul_bin$q_bin_2))/(length(jun_jul_bin$q_bin_2)-length(which(jun_jul_bin$q_bin_0))),
                                       length(which(jun_jul_bin$q_bin_3))/(length(jun_jul_bin$q_bin_3)-length(which(jun_jul_bin$q_bin_0))),
                                       length(which(jun_jul_bin$q_bin_4))/(length(jun_jul_bin$q_bin_4)-length(which(jun_jul_bin$q_bin_0))),
                                       length(which(jun_jul_bin$q_bin_5))/(length(jun_jul_bin$q_bin_5)-length(which(jun_jul_bin$q_bin_0))),
                                       length(which(jun_jul_bin$q_bin_6))/(length(jun_jul_bin$q_bin_6)-length(which(jun_jul_bin$q_bin_0))),
                                       length(which(jun_jul_bin$q_bin_7))/(length(jun_jul_bin$q_bin_7)-length(which(jun_jul_bin$q_bin_0))),
                                       length(which(jun_jul_bin$q_bin_8))/(length(jun_jul_bin$q_bin_8)-length(which(jun_jul_bin$q_bin_0))),
                                       length(which(jun_jul_bin$q_bin_9))/(length(jun_jul_bin$q_bin_9)-length(which(jun_jul_bin$q_bin_0))),
                                       length(which(jun_jul_bin$q_bin_10))/(length(jun_jul_bin$q_bin_10)-length(which(jun_jul_bin$q_bin_0))),
                                       length(which(jun_jul_bin$q_bin_11))/(length(jun_jul_bin$q_bin_11)-length(which(jun_jul_bin$q_bin_0))),
                                       length(which(jun_jul_bin$q_bin_12))/(length(jun_jul_bin$q_bin_12)-length(which(jun_jul_bin$q_bin_0))),
                                       length(which(jun_jul_bin$q_bin_13))/(length(jun_jul_bin$q_bin_13)-length(which(jun_jul_bin$q_bin_0))),
                                       length(which(jun_jul_bin$q_bin_14))/(length(jun_jul_bin$q_bin_14)-length(which(jun_jul_bin$q_bin_0))),
                                       length(which(jun_jul_bin$q_bin_15))/(length(jun_jul_bin$q_bin_15)-length(which(jun_jul_bin$q_bin_0))))

colnames(q_prob_scenario2_jun_jul) <- c("q_bin_1", "q_bin_2","q_bin_3","q_bin_4","q_bin_5","q_bin_6","q_bin_7",
                                    "q_bin_8", "q_bin_9","q_bin_10","q_bin_11","q_bin_12","q_bin_13","q_bin_14","q_bin_15")

q_prob_scenario2_aug_sep <- cbind(length(which(aug_sep_bin$q_bin_1))/(length(aug_sep_bin$q_bin_1)-length(which(aug_sep_bin$q_bin_0))),
                                       length(which(aug_sep_bin$q_bin_2))/(length(aug_sep_bin$q_bin_2)-length(which(aug_sep_bin$q_bin_0))),
                                       length(which(aug_sep_bin$q_bin_3))/(length(aug_sep_bin$q_bin_3)-length(which(aug_sep_bin$q_bin_0))),
                                       length(which(aug_sep_bin$q_bin_4))/(length(aug_sep_bin$q_bin_4)-length(which(aug_sep_bin$q_bin_0))),
                                       length(which(aug_sep_bin$q_bin_5))/(length(aug_sep_bin$q_bin_5)-length(which(aug_sep_bin$q_bin_0))),
                                       length(which(aug_sep_bin$q_bin_6))/(length(aug_sep_bin$q_bin_6)-length(which(aug_sep_bin$q_bin_0))),
                                       length(which(aug_sep_bin$q_bin_7))/(length(aug_sep_bin$q_bin_7)-length(which(aug_sep_bin$q_bin_0))),
                                       length(which(aug_sep_bin$q_bin_8))/(length(aug_sep_bin$q_bin_8)-length(which(aug_sep_bin$q_bin_0))),
                                       length(which(aug_sep_bin$q_bin_9))/(length(aug_sep_bin$q_bin_9)-length(which(aug_sep_bin$q_bin_0))),
                                       length(which(aug_sep_bin$q_bin_10))/(length(aug_sep_bin$q_bin_10)-length(which(aug_sep_bin$q_bin_0))),
                                       length(which(aug_sep_bin$q_bin_11))/(length(aug_sep_bin$q_bin_11)-length(which(aug_sep_bin$q_bin_0))),
                                       length(which(aug_sep_bin$q_bin_12))/(length(aug_sep_bin$q_bin_12)-length(which(aug_sep_bin$q_bin_0))),
                                       length(which(aug_sep_bin$q_bin_13))/(length(aug_sep_bin$q_bin_13)-length(which(aug_sep_bin$q_bin_0))),
                                       length(which(aug_sep_bin$q_bin_14))/(length(aug_sep_bin$q_bin_14)-length(which(aug_sep_bin$q_bin_0))),
                                       length(which(aug_sep_bin$q_bin_15))/(length(aug_sep_bin$q_bin_15)-length(which(aug_sep_bin$q_bin_0))))

colnames(q_prob_scenario2_aug_sep) <- c("q_bin_1", "q_bin_2","q_bin_3","q_bin_4","q_bin_5","q_bin_6","q_bin_7",
                                    "q_bin_8", "q_bin_9","q_bin_10","q_bin_11","q_bin_12","q_bin_13","q_bin_14","q_bin_15")

q_prob_scenario2_discrete <- rbind(q_prob_scenario2_apr_may,q_prob_scenario2_jun_jul,q_prob_scenario2_aug_sep)
row.names(q_prob_scenario2_discrete) <- c("apr_may","jun_jul","aug_sep")
q_prob_scenario2_discrete <- q_prob_scenario2_discrete/rowSums(q_prob_scenario2_discrete)
```


Calculates the probability of a bin flood event occuring for scenario3 flow scenario for discrete time states (April-May, June-July, August-September). Each flow scenario corrects the probability distribution for error by normalizing the probabilities (i.e. ensuring the distribution sums to 1).
```{r}
apr_may_bin <- q_bin_scenario3[months(q_bin_scenario3$date) %in% month.name[4:5],]
jun_jul_bin <- q_bin_scenario3[months(q_bin_scenario3$date) %in% month.name[6:7],]
aug_sep_bin <- q_bin_scenario3[months(q_bin_scenario3$date) %in% month.name[8:9],]

q_prob_scenario3_apr_may <- cbind(length(which(apr_may_bin$q_bin_1))/(length(apr_may_bin$q_bin_1)-length(which(apr_may_bin$q_bin_0))),
                                       length(which(apr_may_bin$q_bin_2))/(length(apr_may_bin$q_bin_2)-length(which(apr_may_bin$q_bin_0))),
                                       length(which(apr_may_bin$q_bin_3))/(length(apr_may_bin$q_bin_3)-length(which(apr_may_bin$q_bin_0))),
                                       length(which(apr_may_bin$q_bin_4))/(length(apr_may_bin$q_bin_4)-length(which(apr_may_bin$q_bin_0))),
                                       length(which(apr_may_bin$q_bin_5))/(length(apr_may_bin$q_bin_5)-length(which(apr_may_bin$q_bin_0))),
                                       length(which(apr_may_bin$q_bin_6))/(length(apr_may_bin$q_bin_6)-length(which(apr_may_bin$q_bin_0))),
                                       length(which(apr_may_bin$q_bin_7))/(length(apr_may_bin$q_bin_7)-length(which(apr_may_bin$q_bin_0))),
                                       length(which(apr_may_bin$q_bin_8))/(length(apr_may_bin$q_bin_8)-length(which(apr_may_bin$q_bin_0))),
                                       length(which(apr_may_bin$q_bin_9))/(length(apr_may_bin$q_bin_9)-length(which(apr_may_bin$q_bin_0))),
                                       length(which(apr_may_bin$q_bin_10))/(length(apr_may_bin$q_bin_10)-length(which(apr_may_bin$q_bin_0))),
                                       length(which(apr_may_bin$q_bin_11))/(length(apr_may_bin$q_bin_11)-length(which(apr_may_bin$q_bin_0))),
                                       length(which(apr_may_bin$q_bin_12))/(length(apr_may_bin$q_bin_12)-length(which(apr_may_bin$q_bin_0))),
                                       length(which(apr_may_bin$q_bin_13))/(length(apr_may_bin$q_bin_13)-length(which(apr_may_bin$q_bin_0))),
                                       length(which(apr_may_bin$q_bin_14))/(length(apr_may_bin$q_bin_14)-length(which(apr_may_bin$q_bin_0))),
                                       length(which(apr_may_bin$q_bin_15))/(length(apr_may_bin$q_bin_15)-length(which(apr_may_bin$q_bin_0))))

colnames(q_prob_scenario3_apr_may) <- c("q_bin_1", "q_bin_2","q_bin_3","q_bin_4","q_bin_5","q_bin_6","q_bin_7",
                                    "q_bin_8", "q_bin_9","q_bin_10","q_bin_11","q_bin_12","q_bin_13","q_bin_14","q_bin_15")

q_prob_scenario3_jun_jul <- cbind(length(which(jun_jul_bin$q_bin_1))/(length(jun_jul_bin$q_bin_1)-length(which(jun_jul_bin$q_bin_0))),
                                       length(which(jun_jul_bin$q_bin_2))/(length(jun_jul_bin$q_bin_2)-length(which(jun_jul_bin$q_bin_0))),
                                       length(which(jun_jul_bin$q_bin_3))/(length(jun_jul_bin$q_bin_3)-length(which(jun_jul_bin$q_bin_0))),
                                       length(which(jun_jul_bin$q_bin_4))/(length(jun_jul_bin$q_bin_4)-length(which(jun_jul_bin$q_bin_0))),
                                       length(which(jun_jul_bin$q_bin_5))/(length(jun_jul_bin$q_bin_5)-length(which(jun_jul_bin$q_bin_0))),
                                       length(which(jun_jul_bin$q_bin_6))/(length(jun_jul_bin$q_bin_6)-length(which(jun_jul_bin$q_bin_0))),
                                       length(which(jun_jul_bin$q_bin_7))/(length(jun_jul_bin$q_bin_7)-length(which(jun_jul_bin$q_bin_0))),
                                       length(which(jun_jul_bin$q_bin_8))/(length(jun_jul_bin$q_bin_8)-length(which(jun_jul_bin$q_bin_0))),
                                       length(which(jun_jul_bin$q_bin_9))/(length(jun_jul_bin$q_bin_9)-length(which(jun_jul_bin$q_bin_0))),
                                       length(which(jun_jul_bin$q_bin_10))/(length(jun_jul_bin$q_bin_10)-length(which(jun_jul_bin$q_bin_0))),
                                       length(which(jun_jul_bin$q_bin_11))/(length(jun_jul_bin$q_bin_11)-length(which(jun_jul_bin$q_bin_0))),
                                       length(which(jun_jul_bin$q_bin_12))/(length(jun_jul_bin$q_bin_12)-length(which(jun_jul_bin$q_bin_0))),
                                       length(which(jun_jul_bin$q_bin_13))/(length(jun_jul_bin$q_bin_13)-length(which(jun_jul_bin$q_bin_0))),
                                       length(which(jun_jul_bin$q_bin_14))/(length(jun_jul_bin$q_bin_14)-length(which(jun_jul_bin$q_bin_0))),
                                       length(which(jun_jul_bin$q_bin_15))/(length(jun_jul_bin$q_bin_15)-length(which(jun_jul_bin$q_bin_0))))

colnames(q_prob_scenario3_jun_jul) <- c("q_bin_1", "q_bin_2","q_bin_3","q_bin_4","q_bin_5","q_bin_6","q_bin_7",
                                    "q_bin_8", "q_bin_9","q_bin_10","q_bin_11","q_bin_12","q_bin_13","q_bin_14","q_bin_15")

q_prob_scenario3_aug_sep <- cbind(length(which(aug_sep_bin$q_bin_1))/(length(aug_sep_bin$q_bin_1)-length(which(aug_sep_bin$q_bin_0))),
                                       length(which(aug_sep_bin$q_bin_2))/(length(aug_sep_bin$q_bin_2)-length(which(aug_sep_bin$q_bin_0))),
                                       length(which(aug_sep_bin$q_bin_3))/(length(aug_sep_bin$q_bin_3)-length(which(aug_sep_bin$q_bin_0))),
                                       length(which(aug_sep_bin$q_bin_4))/(length(aug_sep_bin$q_bin_4)-length(which(aug_sep_bin$q_bin_0))),
                                       length(which(aug_sep_bin$q_bin_5))/(length(aug_sep_bin$q_bin_5)-length(which(aug_sep_bin$q_bin_0))),
                                       length(which(aug_sep_bin$q_bin_6))/(length(aug_sep_bin$q_bin_6)-length(which(aug_sep_bin$q_bin_0))),
                                       length(which(aug_sep_bin$q_bin_7))/(length(aug_sep_bin$q_bin_7)-length(which(aug_sep_bin$q_bin_0))),
                                       length(which(aug_sep_bin$q_bin_8))/(length(aug_sep_bin$q_bin_8)-length(which(aug_sep_bin$q_bin_0))),
                                       length(which(aug_sep_bin$q_bin_9))/(length(aug_sep_bin$q_bin_9)-length(which(aug_sep_bin$q_bin_0))),
                                       length(which(aug_sep_bin$q_bin_10))/(length(aug_sep_bin$q_bin_10)-length(which(aug_sep_bin$q_bin_0))),
                                       length(which(aug_sep_bin$q_bin_11))/(length(aug_sep_bin$q_bin_11)-length(which(aug_sep_bin$q_bin_0))),
                                       length(which(aug_sep_bin$q_bin_12))/(length(aug_sep_bin$q_bin_12)-length(which(aug_sep_bin$q_bin_0))),
                                       length(which(aug_sep_bin$q_bin_13))/(length(aug_sep_bin$q_bin_13)-length(which(aug_sep_bin$q_bin_0))),
                                       length(which(aug_sep_bin$q_bin_14))/(length(aug_sep_bin$q_bin_14)-length(which(aug_sep_bin$q_bin_0))),
                                       length(which(aug_sep_bin$q_bin_15))/(length(aug_sep_bin$q_bin_15)-length(which(aug_sep_bin$q_bin_0))))

colnames(q_prob_scenario3_aug_sep) <- c("q_bin_1", "q_bin_2","q_bin_3","q_bin_4","q_bin_5","q_bin_6","q_bin_7",
                                    "q_bin_8", "q_bin_9","q_bin_10","q_bin_11","q_bin_12","q_bin_13","q_bin_14","q_bin_15")

q_prob_scenario3_discrete <- rbind(q_prob_scenario3_apr_may,q_prob_scenario3_jun_jul,q_prob_scenario3_aug_sep)
row.names(q_prob_scenario3_discrete) <- c("apr_may","jun_jul","aug_sep")
q_prob_scenario3_discrete <- q_prob_scenario3_discrete/rowSums(q_prob_scenario3_discrete)
```
