==> 2-gila-BNstructure_new.Rmd <==
---
title: Upper Gila Watershed Bayesian Network Structure
author: Lauren Victoria Jaramillo, UNM (alwayslo@unm.edu)
date: 1/8/2019
update: 8/20/2019
output: html_notebook
---

This R notebook is used to used to build and Bayesian network used to infer recruitment potential. Conditional probabilities for each node are populated using results from the "1-gila-preprocessing.Rmd".**Bayesian networks** are a class of graphical models that allow a concise representation of the probabilistic dependencies between a given set of random variables as a **directed acyclic graph** where each node represents a random variable.

###Model Development###
This first code chunck clears the environment, installs all the necessary packages, and returns the current working directory to ensure filespaths will execute properly.This code chunck loads and attaches the add-on packages *bnlearn*,*readr*,and *Rgraphviz*.*Rgraphviz* requires additional packages *BioGenerics* and graph to run  It also clears the the global environment of any previous data.
```{r}
#rm(list=ls())

library("bnlearn")
library("readr")
library("Rgraphviz")
library("xts")
library("prob")
library("TTR")
library("bigmemory")
library("zoo")
library("raster")
library("sp")
library("rgdal")
library("rasterVis")
library("ggplot2")
library("rgdal")
library("methods")
require("BiocGenerics")
require("graph")

getwd()
```

####Creating Network Structure#### 
There are several approaches to creating a network structure in R, this code chunck creates a new bn object by creating an empty graph using using our variables established in dag and directed arcs using by using *set.arc*. These directed arcs can be modified if expert knowedge suggests a different relationship using *drop.arc*,and *rev.arc*.

Below generates the network structure as presented in Figure 4 of 2014 Morrison et al.
```{r}
nodes <- c("TIMING","INUN","RECESSION_RATE","HYDRO","GW","RECRUIT")

dag = empty.graph(nodes,num=1)
dag = set.arc(dag, "TIMING", "INUN", check.cycles = TRUE, check.illegal = TRUE, debug = FALSE)
dag = set.arc(dag, "TIMING", "RECESSION_RATE", check.cycles = TRUE, check.illegal = TRUE, debug = FALSE)
dag = set.arc(dag, "INUN", "HYDRO", check.cycles = TRUE, check.illegal = TRUE, debug = FALSE)
dag = set.arc(dag, "RECESSION_RATE", "HYDRO", check.cycles = TRUE, check.illegal = TRUE, debug = FALSE)
dag = set.arc(dag, "GW", "RECRUIT", check.cycles = TRUE, check.illegal = TRUE, debug = FALSE)
dag = set.arc(dag, "HYDRO", "RECRUIT", check.cycles = TRUE, check.illegal = TRUE, debug = FALSE)
```

###Plotting Network Structure: Advanced Bayesian Network Plot###
Visualization of the generated graph allows us to view and confirm the relationships (dependencies, direction, children, parents, etc).The following code chunck plots the network structure using *graphviz.plot*. The graph layout options for *graphviz.plot* include dot, neato, twopi, circo, and fdp.
```{r}

graphviz.plot(dag, highlight = NULL, layout = "dot",shape = "rectangle", main = "Gila Bayesian Network Structure", sub = NULL)

```
###Topological Ordering and Investigating the Properities of a BN###

The topological ordering of the nodes is important to this project specifically because it is a causal model. The ordering can be determined using *node.ordering*.The neighborhood *nbr* and Markov blanket *mb* of a node is a synthetic discription of the local dependence structure.

*parents* and *children* are used to show the children and parents of a particular node given the established network structure.*vstructs* show the series of converging connections in the network. The converging connections are displayed in a matrix and the nodes listed column X and Y are converging towardsthe node listed in Z column. 

*cextend* prints a summary of the network including number of nodes, number of arcs (undirected and directed), the average markov blanket size, average neighborhood size, and average branching factor. In any BN, the markov blanket of a node is the set of parents of that node, the children of that node, and all other nodes sharing a child with that node. The neighborhood size is the number of adjacent nodes. The branching factor is the number of children for a node in the network.

*directed.arcs* is a way of verifying which arcs are directed and from "" to "". 
```{r}
node.ordering(dag)

nbr(dag, "TIMING")
mb(dag, "TIMING")

nbr(dag, "INUN")
nbr(dag, "GW")

mb(dag, "INUN")
mb(dag, "GW")

parents(dag, "GW")
parents(dag, "INUN")

children(dag, "GW")
children(dag, "INUN")

vstructs(dag, moral = TRUE)

cextend(dag)

directed.arcs(dag)
```

#### Discrete States ####
The discrete mutually exclusice values or states for each BN node is established below which is part of the discrete static Bayesian network infered from expert knowledge and literature. 

#### CPT Assignments ####
The unconditional and conditional probability tables are generated below cased on the scenario specified and run. The unconditional probability table (upt) for the root node, TIMING, was developed based on the historical dataset (need to verify). The CPTs for each node must share the levels of its parents. The number of elements needed for the array/matrix must be equal the product of the number of states of the parent nodes and the current node.

cpt_HYDRO was populated using the expert knowledge and informed from the probabilities from 2014 Morrison & Stone.

The conditional probability tables were developed in "1-gila-preprocessing.Rmd" and are presented here.

#### Natural ####
```{r}
upt_TIMING = matrix(data = c(0.65,.25,0.1), nrow = 1 , ncol = 3, dimnames = list(NULL,'TIMING'=c("apr_may", "jun_jul", "aug_sep")))

cpt_INUN = t(matrix(data = c(q_prob_natural_discrete) , nrow = 3 , ncol = 15, dimnames = list('TIMING'=c("apr_may", "jun_jul", "aug_sep"), 'INUN'=c("bin_1", "bin_2","bin_3","bin_4","bin_5","bin_6","bin_7","bin_8", "bin_9","bin_10","bin_11","bin_12","bin_13","bin_14","bin_15"))))

cpt_RECESSION_RATE = t(matrix(data = c(rr_prob_natural_discrete) , nrow = 3 , ncol = 5, dimnames = list('TIMING'=c("apr_may", "jun_jul", "aug_sep"), 'RECESSION_RATE'=c("bin_1", "bin_2", "bin_3", "bin_4", "bin_5"))))

cpt_GW  = matrix(c(0.33,0.33,0.34), nrow = 1, ncol = 3, dimnames = list(NULL,c("<50cm", "50-200cm", "> 200cm")))

cpt_HYDRO = array(data = c(1,0,1,0,1,0,1,0,1,0,1,0,1,0,1,0,1,0,1,0,1,0,1,0,1,0,1,0,1,0,
                           
                           0.95,0.05,0.95,0.05,0.95,0.05,0.95,0.05,0.95,0.05,0.95,0.05,0.95,0.05,0.95,0.05,
                           0.95,0.05,0.95,0.05,0.95,0.05,0.95,0.05,0.95,0.05,0.95,0.05,0.95,0.05,
                           
                           0.1,0.9,0.1,0.9,0.1,0.9,0.1,0.9,0.1,0.9,0.1,0.9,0.1,0.9,0.1,0.9,
                           0.1,0.9,0.1,0.9,0.1,0.9,0.1,0.9,0.1,0.9,0.1,0.9,0.1,0.9,
                           
                           0.3,0.7,0.3,0.7,0.3,0.7,0.3,0.7,0.3,0.7,0.3,0.7,0.3,0.7,0.3,0.7,
                           0.3,0.7,0.3,0.7,0.3,0.7,0.3,0.7,0.3,0.7,0.3,0.7,0.3,0.7,
                           
                           0.9,0.1, 0.9,0.1,0.9,0.1,0.9,0.1,0.9,0.1,0.9,0.1,0.9,0.1,0.9,0.1,
                           0.9,0.1, 0.9,0.1,0.9,0.1,0.9,0.1,0.9,0.1,0.9,0.1,0.9,0.1),
              dim = c(2,15,5), 
              dimnames =list('HYDRO'= c("LOW","HIGH"),
                           'INUN' = c("bin_1", "bin_2","bin_3","bin_4","bin_5","bin_6","bin_7",
                                    "bin_8", "bin_9","bin_10","bin_11","bin_12","bin_13","bin_14","bin_15"),
                           'RECESSION_RATE'= c("bin_1", "bin_2", "bin_3", "bin_4", "bin_5")))
  
cpt_RECRUIT = array(data=c(0.5, 0.5, 0.7, 0.3, 0.5, 0.5, 0.9, 0.1, 0.5, 0.5, 0.7, 0.3), 
              dim= c(2,2,3),  
              dimnames = list('RECRUIT'=c("YES","NO"),
                              'HYDRO'= c("LOW","HIGH"),
                              'GW'= c("<50cm", "50-200cm", "> 200cm")))
```

#### Scenario 1 ####
```{r}
upt_TIMING = matrix(data = c(0.65,.25,0.1), nrow = 1 , ncol = 3, dimnames = list(NULL,'TIMING'=c("apr_may", "jun_jul", "aug_sep")))

cpt_INUN = t(matrix(data = c(q_prob_scenario1_discrete) , nrow = 3 , ncol = 15, dimnames = list('TIMING'=c("apr_may", "jun_jul", "aug_sep"),
          'INUN'=c("bin_1", "bin_2","bin_3","bin_4","bin_5","bin_6","bin_7","bin_8","bin_9",
                   "bin_10","bin_11","bin_12","bin_13","bin_14","bin_15"))))

cpt_RECESSION_RATE = t(matrix(data = c(rr_prob_scenario1_discrete) , nrow = 3 , ncol = 5, dimnames = list('TIMING'=c("apr_may", "jun_jul", "aug_sep"), 'RECESSION_RATE'=c("bin_1", "bin_2", "bin_3", "bin_4", "bin_5"))))

cpt_GW  = matrix(c(0.33,0.33,0.34), nrow = 1, ncol = 3, dimnames = list(NULL,c("<50cm", "50-200cm", "> 200cm")))

cpt_HYDRO = array(data = c(1,0,1,0,1,0,1,0,1,0,1,0,1,0,1,0,1,0,1,0,1,0,1,0,1,0,1,0,1,0,
                           
                           0.95,0.05,0.95,0.05,0.95,0.05,0.95,0.05,0.95,0.05,0.95,0.05,0.95,0.05,0.95,0.05,
                           0.95,0.05,0.95,0.05,0.95,0.05,0.95,0.05,0.95,0.05,0.95,0.05,0.95,0.05,
                           
                           0.1,0.9,0.1,0.9,0.1,0.9,0.1,0.9,0.1,0.9,0.1,0.9,0.1,0.9,0.1,0.9,
                           0.1,0.9,0.1,0.9,0.1,0.9,0.1,0.9,0.1,0.9,0.1,0.9,0.1,0.9,
                           
                           0.3,0.7,0.3,0.7,0.3,0.7,0.3,0.7,0.3,0.7,0.3,0.7,0.3,0.7,0.3,0.7,
                           0.3,0.7,0.3,0.7,0.3,0.7,0.3,0.7,0.3,0.7,0.3,0.7,0.3,0.7,
                           
                           0.9,0.1, 0.9,0.1,0.9,0.1,0.9,0.1,0.9,0.1,0.9,0.1,0.9,0.1,0.9,0.1,
                           0.9,0.1, 0.9,0.1,0.9,0.1,0.9,0.1,0.9,0.1,0.9,0.1,0.9,0.1),
              dim = c(2,15,5), 
              dimnames =list('HYDRO'= c("LOW","HIGH"),
                           'INUN' = c("bin_1", "bin_2","bin_3","bin_4","bin_5","bin_6","bin_7",
                                    "bin_8", "bin_9","bin_10","bin_11","bin_12","bin_13","bin_14","bin_15"),
                           'RECESSION_RATE'= c("bin_1", "bin_2", "bin_3", "bin_4", "bin_5")))
  
cpt_RECRUIT = array(data=c(0.5, 0.5, 0.7, 0.3, 0.5, 0.5, 0.9, 0.1, 0.5, 0.5, 0.7, 0.3), 
              dim= c(2,2,3),  
              dimnames = list('RECRUIT'=c("YES","NO"),
                              'HYDRO'= c("LOW","HIGH"),
                              'GW'= c("<50cm", "50-200cm", "> 200cm")))
```


#### Scenario 2 ####
```{r}
upt_TIMING = matrix(data = c(0.65,.25,0.1), nrow = 1 , ncol = 3, dimnames = list(NULL,'TIMING'=c("apr_may", "jun_jul", "aug_sep")))

cpt_INUN = t(matrix(data = c(q_prob_scenario2_discrete) , nrow = 3 , ncol = 15, dimnames = list('TIMING'=c("apr_may", "jun_jul", "aug_sep"),
          'INUN'=c("bin_1", "bin_2","bin_3","bin_4","bin_5","bin_6","bin_7","bin_8","bin_9",
                   "bin_10","bin_11","bin_12","bin_13","bin_14","bin_15"))))

cpt_RECESSION_RATE = t(matrix(data = c(rr_prob_scenario2_discrete) , nrow = 3 , ncol = 5, dimnames = list('TIMING'=c("apr_may", "jun_jul", "aug_sep"), 'RECESSION_RATE'=c("bin_1", "bin_2", "bin_3", "bin_4", "bin_5"))))

cpt_GW  = matrix(c(0.33,0.33,0.34), nrow = 1, ncol = 3, dimnames = list(NULL,c("<50cm", "50-200cm", "> 200cm")))

cpt_HYDRO = array(data = c(1,0,1,0,1,0,1,0,1,0,1,0,1,0,1,0,1,0,1,0,1,0,1,0,1,0,1,0,1,0,
                           
                           0.95,0.05,0.95,0.05,0.95,0.05,0.95,0.05,0.95,0.05,0.95,0.05,0.95,0.05,0.95,0.05,
                           0.95,0.05,0.95,0.05,0.95,0.05,0.95,0.05,0.95,0.05,0.95,0.05,0.95,0.05,
                           
                           0.1,0.9,0.1,0.9,0.1,0.9,0.1,0.9,0.1,0.9,0.1,0.9,0.1,0.9,0.1,0.9,
                           0.1,0.9,0.1,0.9,0.1,0.9,0.1,0.9,0.1,0.9,0.1,0.9,0.1,0.9,
                           
                           0.3,0.7,0.3,0.7,0.3,0.7,0.3,0.7,0.3,0.7,0.3,0.7,0.3,0.7,0.3,0.7,
                           0.3,0.7,0.3,0.7,0.3,0.7,0.3,0.7,0.3,0.7,0.3,0.7,0.3,0.7,
                           
                           0.9,0.1, 0.9,0.1,0.9,0.1,0.9,0.1,0.9,0.1,0.9,0.1,0.9,0.1,0.9,0.1,
                           0.9,0.1, 0.9,0.1,0.9,0.1,0.9,0.1,0.9,0.1,0.9,0.1,0.9,0.1),
              dim = c(2,15,5), 
              dimnames =list('HYDRO'= c("LOW","HIGH"),
                           'INUN' = c("bin_1", "bin_2","bin_3","bin_4","bin_5","bin_6","bin_7",
                                    "bin_8", "bin_9","bin_10","bin_11","bin_12","bin_13","bin_14","bin_15"),
                           'RECESSION_RATE'= c("bin_1", "bin_2", "bin_3", "bin_4", "bin_5")))
  
cpt_RECRUIT = array(data=c(0.5, 0.5, 0.7, 0.3, 0.5, 0.5, 0.9, 0.1, 0.5, 0.5, 0.7, 0.3), 
              dim= c(2,2,3),  
              dimnames = list('RECRUIT'=c("YES","NO"),
                              'HYDRO'= c("LOW","HIGH"),
                              'GW'= c("<50cm", "50-200cm", "> 200cm")))
```

#### Scenario 3 ####
```{r}
upt_TIMING = matrix(data = c(0.65,.25,0.1), nrow = 1 , ncol = 3, dimnames = list(NULL,'TIMING'=c("apr_may", "jun_jul", "aug_sep")))

cpt_INUN = t(matrix(data = c(q_prob_scenario3_discrete) , nrow = 3 , ncol = 15, dimnames = list('TIMING'=c("apr_may", "jun_jul", "aug_sep"), 
          'INUN'=c("bin_1", "bin_2","bin_3","bin_4","bin_5","bin_6","bin_7","bin_8","bin_9",
                   "bin_10","bin_11","bin_12","bin_13","bin_14","bin_15"))))

cpt_RECESSION_RATE = t(matrix(data = (rr_prob_scenario3_discrete) , nrow = 3 , ncol = 5, dimnames = list('TIMING'=c("apr_may", "jun_jul", "aug_sep"), 'RECESSION_RATE'=c("bin_1", "bin_2", "bin_3", "bin_4", "bin_5"))))

cpt_GW  = matrix(c(0.33,0.33,0.34), nrow = 1, ncol = 3, dimnames = list(NULL,c("<50cm", "50-200cm", "> 200cm")))

cpt_HYDRO = array(data = c(1,0,1,0,1,0,1,0,1,0,1,0,1,0,1,0,1,0,1,0,1,0,1,0,1,0,1,0,1,0,
                           
                           0.95,0.05,0.95,0.05,0.95,0.05,0.95,0.05,0.95,0.05,0.95,0.05,0.95,0.05,0.95,0.05,
                           0.95,0.05,0.95,0.05,0.95,0.05,0.95,0.05,0.95,0.05,0.95,0.05,0.95,0.05,
                           
                           0.1,0.9,0.1,0.9,0.1,0.9,0.1,0.9,0.1,0.9,0.1,0.9,0.1,0.9,0.1,0.9,
                           0.1,0.9,0.1,0.9,0.1,0.9,0.1,0.9,0.1,0.9,0.1,0.9,0.1,0.9,
                           
                           0.3,0.7,0.3,0.7,0.3,0.7,0.3,0.7,0.3,0.7,0.3,0.7,0.3,0.7,0.3,0.7,
                           0.3,0.7,0.3,0.7,0.3,0.7,0.3,0.7,0.3,0.7,0.3,0.7,0.3,0.7,
                           
                           0.9,0.1, 0.9,0.1,0.9,0.1,0.9,0.1,0.9,0.1,0.9,0.1,0.9,0.1,0.9,0.1,
                           0.9,0.1, 0.9,0.1,0.9,0.1,0.9,0.1,0.9,0.1,0.9,0.1,0.9,0.1),
              dim = c(2,15,5), 
              dimnames =list('HYDRO'= c("LOW","HIGH"),
                           'INUN' = c("bin_1", "bin_2","bin_3","bin_4","bin_5","bin_6","bin_7",
                                    "bin_8", "bin_9","bin_10","bin_11","bin_12","bin_13","bin_14","bin_15"),
                           'RECESSION_RATE'= c("bin_1", "bin_2", "bin_3", "bin_4", "bin_5")))
  
cpt_RECRUIT = array(data=c(0.5, 0.5, 0.7, 0.3, 0.5, 0.5, 0.9, 0.1, 0.5, 0.5, 0.7, 0.3), 
              dim= c(2,2,3),  
              dimnames = list('RECRUIT'=c("YES","NO"),
                              'HYDRO'= c("LOW","HIGH"),
                              'GW'= c("<50cm", "50-200cm", "> 200cm")))
```

The following assigns the established conditional probability tables (CPTs) to each node.
```{r}
bnfit = custom.fit(dag,dist = list (TIMING = upt_TIMING,
                                    INUN = cpt_INUN,
                                    RECESSION_RATE = cpt_RECESSION_RATE,
                                    GW = cpt_GW,
                                    HYDRO=cpt_HYDRO,
                                    RECRUIT = cpt_RECRUIT))

 cpquery(bnfit,event = (RECRUIT =="YES"),evidence = list(TIMING="aug_sep",INUN="bin_1",RECESSION_RATE = "bin_1"), method = "lw")
```


###References ###
Gentleman R, Whalen E, Huber W and Falcon S (2017). graph: graph: A package to handle graph data structures. R package version 1.54.0.

Hansen KD, Gentry J, Long L, Gentleman R, Falcon S, Hahne F and Sarkar D (2017). Rgraphviz: Provides plotting capabilities for R graph objects. R package version 2.20.0.

Huber, W., Carey, J. V, Gentleman, R., Anders, S., Carlson, M., Carvalho, S. B, Bravo, C. H, Davis, S., Gatto, L., Girke, T., Gottardo, R., Hahne, F., Hansen, D. K, Irizarry, A. R, Lawrence, M., Love, I. M, MacDonald, J., Obenchain, V., Ole's, K. A, Pag'es, H., Reyes, A., Shannon, P., Smyth, K. G, Tenenbaum, D., Waldron, L., Morgan and M. (2015). “Orchestrating high-throughput genomic analysis with Bioconductor.” Nature Methods, 12(2), pp. 115–121. http://www.nature.com/nmeth/journal/v12/n2/full/nmeth.3252.html.

Koller D, Friedman N (2009). Probabilistic Graphical Models: Principles and Techniques. MIT Press.

Pearl J (2009). Causality: Models, Reasoning and Inference. Cambridge University Press, 2nd edition.

Nagarajan, Radhakrishnan, Marco Scutari, and Sophie Lèbre. Bayesian Networks in R. New York, NY: Springer New York, 2013. http://link.springer.com/10.1007/978-1-4614-6446-4. 
http://www.bnlearn.com

Scutari, Marco (2018). bnlearn: Bayesian Network Structure Learning, Parameter Learning and Inference. R package version 4.4 https://cran.r-project.org/web/packages/bnlearn/

https://stackoverflow.com/questions/40364458/r-some-conditional-probability-distributions-of-node-do-not-sum-to-one

http://www.bnlearn.com/bnrepository/discrete-small.html#earthquake

source code - https://rdrr.io/cran/bnlearn/src/R/utils-cluster.R#sym-smartSapply

cpquery - http://www.bnlearn.com/documentation/man/cpquery.html
