==> 3-gila-implement_new.Rmd <==
---
title: Upper Gila Watershed Bayesian Network Implementation
author: Lauren Victoria Jaramillo, UNM (alwayslo@unm.edu)
date: 4/2/2019
update: 8/21/2019
output: html_notebook
---

This R notebook is used to used to implement the constructed Bayesian network for riparian recruitment for various scenarios (natural, scenario 1, scenario 2,and scenario 3). These scenarios corresponde to the new CUFA.

This first code chunck clears the environment, installs all the necessary packages, and returns the current working directory to ensure filespaths will execute properly.This code chunck loads and attaches the add-on packages *bnlearn*,*readr*,and *Rgraphviz*.*Rgraphviz* requires additional packages *BioGenerics* and graph to run  It also clears the the global environment of any previous data.
```{r}
#rm(list=ls())

library("bnlearn")
library("readr")
library("Rgraphviz")
library("xts")
library("prob")
library("TTR")
library("bigmemory")
library("zoo")
library("raster")
library("sp")
library("rgdal")
library("rasterVis")
library("ggplot2")
library("rgdal")
library("methods")
require("BiocGenerics")
require("graph")

getwd()
```

###Import SRH-2D model results###
This code chunck calls the txt file that includes the inundation stats for each inundated cell in the hydrodynamic model mesh. This calls the data for one of 3 sites (1-North, 2-Middle, 3-South).

<!-- ####Site 1##### -->
<!-- ```{r} -->
<!-- q_cells <- read.table("data/SiteN_Qs.txt",skip=1,header = TRUE) -->
<!-- q_cells <- setNames(q_cells, c("cell","flow")) -->

<!-- q_cells$bin <-      ifelse(q_bin[1]<= q_cells$flow  & q_cells$flow <q_bin[2],"bin_1", -->
<!--                     ifelse(q_bin[2]<= q_cells$flow  & q_cells$flow <q_bin[3],"bin_2", -->
<!--                     ifelse(q_bin[3]<= q_cells$flow  & q_cells$flow <q_bin[4],"bin_3", -->
<!--                     ifelse(q_bin[4]<= q_cells$flow  & q_cells$flow <q_bin[5],"bin_4", -->
<!--                     ifelse(q_bin[5]<= q_cells$flow  & q_cells$flow <q_bin[6],"bin_5", -->
<!--                     ifelse(q_bin[6]<= q_cells$flow  & q_cells$flow <q_bin[7],"bin_6", -->
<!--                     ifelse(q_bin[7]<= q_cells$flow  & q_cells$flow <q_bin[8],"bin_7", -->
<!--                     ifelse(q_bin[8]<= q_cells$flow  & q_cells$flow <q_bin[9],"bin_8", -->
<!--                     ifelse(q_bin[9]<= q_cells$flow  & q_cells$flow <q_bin[10],"bin_9", -->
<!--                     ifelse(q_bin[10]<= q_cells$flow  & q_cells$flow <q_bin[11],"bin_10", -->
<!--                     ifelse(q_bin[11]<= q_cells$flow  & q_cells$flow <q_bin[12],"bin_11", -->
<!--                     ifelse(q_bin[12]<= q_cells$flow  & q_cells$flow <q_bin[13],"bin_12", -->
<!--                     ifelse(q_bin[13]<= q_cells$flow  & q_cells$flow <q_bin[14],"bin_13", -->
<!--                     ifelse(q_bin[14]<= q_cells$flow  & q_cells$flow <q_bin[15],"bin_14", -->
<!--                     ifelse(q_bin[15]<= q_cells$flow,"bin_15","blah"))))))))))))))) -->

<!-- ``` -->

####Site 2#####
```{r}
q_cells <- read.table("data/SiteM_Qs.txt",skip=1,header = TRUE)
q_cells <- setNames(q_cells, c("cell","flow"))

q_cells$bin <-      ifelse(q_bin[1]<= q_cells$flow  & q_cells$flow <q_bin[2],"bin_1",
                    ifelse(q_bin[2]<= q_cells$flow  & q_cells$flow <q_bin[3],"bin_2",
                    ifelse(q_bin[3]<= q_cells$flow  & q_cells$flow <q_bin[4],"bin_3",
                    ifelse(q_bin[4]<= q_cells$flow  & q_cells$flow <q_bin[5],"bin_4",
                    ifelse(q_bin[5]<= q_cells$flow  & q_cells$flow <q_bin[6],"bin_5",
                    ifelse(q_bin[6]<= q_cells$flow  & q_cells$flow <q_bin[7],"bin_6",
                    ifelse(q_bin[7]<= q_cells$flow  & q_cells$flow <q_bin[8],"bin_7",
                    ifelse(q_bin[8]<= q_cells$flow  & q_cells$flow <q_bin[9],"bin_8",
                    ifelse(q_bin[9]<= q_cells$flow  & q_cells$flow <q_bin[10],"bin_9",
                    ifelse(q_bin[10]<= q_cells$flow  & q_cells$flow <q_bin[11],"bin_10",
                    ifelse(q_bin[11]<= q_cells$flow  & q_cells$flow <q_bin[12],"bin_11",
                    ifelse(q_bin[12]<= q_cells$flow  & q_cells$flow <q_bin[13],"bin_12",
                    ifelse(q_bin[13]<= q_cells$flow  & q_cells$flow <q_bin[14],"bin_13",
                    ifelse(q_bin[14]<= q_cells$flow  & q_cells$flow <q_bin[15],"bin_14",
                    ifelse(q_bin[15]<= q_cells$flow,"bin_15","blah")))))))))))))))

```

<!-- ####Site 3##### -->
<!-- ```{r} -->
<!-- q_cells <- read.table("data/SiteS_Qs.txt",skip=1,header = TRUE) -->
<!-- q_cells <- setNames(q_cells, c("cell","flow")) -->

<!-- q_cells$bin <-      ifelse(q_bin[1]<= q_cells$flow  & q_cells$flow <q_bin[2],"bin_1", -->
<!--                     ifelse(q_bin[2]<= q_cells$flow  & q_cells$flow <q_bin[3],"bin_2", -->
<!--                     ifelse(q_bin[3]<= q_cells$flow  & q_cells$flow <q_bin[4],"bin_3", -->
<!--                     ifelse(q_bin[4]<= q_cells$flow  & q_cells$flow <q_bin[5],"bin_4", -->
<!--                     ifelse(q_bin[5]<= q_cells$flow  & q_cells$flow <q_bin[6],"bin_5", -->
<!--                     ifelse(q_bin[6]<= q_cells$flow  & q_cells$flow <q_bin[7],"bin_6", -->
<!--                     ifelse(q_bin[7]<= q_cells$flow  & q_cells$flow <q_bin[8],"bin_7", -->
<!--                     ifelse(q_bin[8]<= q_cells$flow  & q_cells$flow <q_bin[9],"bin_8", -->
<!--                     ifelse(q_bin[9]<= q_cells$flow  & q_cells$flow <q_bin[10],"bin_9", -->
<!--                     ifelse(q_bin[10]<= q_cells$flow  & q_cells$flow <q_bin[11],"bin_10", -->
<!--                     ifelse(q_bin[11]<= q_cells$flow  & q_cells$flow <q_bin[12],"bin_11", -->
<!--                     ifelse(q_bin[12]<= q_cells$flow  & q_cells$flow <q_bin[13],"bin_12", -->
<!--                     ifelse(q_bin[13]<= q_cells$flow  & q_cells$flow <q_bin[14],"bin_13", -->
<!--                     ifelse(q_bin[14]<= q_cells$flow  & q_cells$flow <q_bin[15],"bin_14", -->
<!--                     ifelse(q_bin[15]<= q_cells$flow,"bin_15","blah"))))))))))))))) -->

<!-- ``` -->

####Natural Scenario Evidence####
```{r}
Evidence <- c(hydro[1:nrow(recessionrate_14d),],recessionrate_14d[,2:5])
Evidence <-data.frame(Evidence)
Evidence$TIMING <- NA
Evidence$INUN <- NA
Evidence$RECESSION_RATE<- NA
Evidence <- setNames(Evidence, c("date","q_natural", "q_scenario1","q_scenario2","q_scenario3","rr_natural", "rr_scenario1","rr_scenario2","rr_scenario3", "TIMING","INUN","RECESSION_RATE"))

Evidence$TIMING <-  ifelse(months.Date(Evidence$date) %in% month.name[4:5],"apr_may",
                    ifelse(months.Date(Evidence$date) %in% month.name[6:7],"jun_jul",
                    ifelse(months.Date(Evidence$date) %in% month.name[8:9],"aug_sep","blah")))

Evidence$INUN <-    ifelse(q_bin[1]<= Evidence$q_natural & Evidence$q_natural<q_bin[2],"bin_1",
                    ifelse(q_bin[2]<= Evidence$q_natural & Evidence$q_natural<q_bin[3],"bin_2",
                    ifelse(q_bin[3]<= Evidence$q_natural & Evidence$q_natural<q_bin[4],"bin_3",
                    ifelse(q_bin[4]<= Evidence$q_natural & Evidence$q_natural<q_bin[5],"bin_4",
                    ifelse(q_bin[5]<= Evidence$q_natural & Evidence$q_natural<q_bin[6],"bin_5",
                    ifelse(q_bin[6]<= Evidence$q_natural & Evidence$q_natural<q_bin[7],"bin_6",
                    ifelse(q_bin[7]<= Evidence$q_natural & Evidence$q_natural<q_bin[8],"bin_7",
                    ifelse(q_bin[8]<= Evidence$q_natural & Evidence$q_natural<q_bin[9],"bin_8",
                    ifelse(q_bin[9]<= Evidence$q_natural & Evidence$q_natural<q_bin[10],"bin_9",
                    ifelse(q_bin[10]<= Evidence$q_natural & Evidence$q_natural<q_bin[11],"bin_10",
                    ifelse(q_bin[11]<= Evidence$q_natural & Evidence$q_natural<q_bin[12],"bin_11",
                    ifelse(q_bin[12]<= Evidence$q_natural & Evidence$q_natural<q_bin[13],"bin_12",
                    ifelse(q_bin[13]<= Evidence$q_natural & Evidence$q_natural<q_bin[14],"bin_13",
                    ifelse(q_bin[14]<= Evidence$q_natural & Evidence$q_natural<q_bin[15],"bin_14",
                    ifelse(q_bin[15]<= Evidence$q_natural,"bin_15","blah")))))))))))))))


Evidence$RECESSION_RATE <- ifelse(Evidence$rr_natural <= recessionrate_bin[1],"bin_1",
                           ifelse(recessionrate_bin[1]< Evidence$rr_natural & Evidence$rr_natural <=recessionrate_bin[2],"bin_2",
                           ifelse(recessionrate_bin[2]< Evidence$rr_natural & Evidence$rr_natural <=recessionrate_bin[3],"bin_3",
                           ifelse(recessionrate_bin[3]< Evidence$rr_natural & Evidence$rr_natural <=recessionrate_bin[4],"bin_4",
                           ifelse(Evidence$rr_natural > recessionrate_bin[4],"bin_5","blah")))))  

Evidence <- Evidence[-1:-9]
```

####Scenario 1 Evidence####
```{r}
Evidence <- c(hydro[1:nrow(recessionrate_14d),],recessionrate_14d[,2:5])
Evidence <-data.frame(Evidence)
Evidence$TIMING <- NA
Evidence$INUN <- NA
Evidence$RECESSION_RATE<- NA
Evidence <- setNames(Evidence, c("date","q_natural", "q_scenario1","q_scenario2","q_scenario3","rr_natural", "rr_scenario1","rr_scenario2","rr_scenario3", "TIMING","INUN","RECESSION_RATE"))

Evidence$TIMING <-  ifelse(months(Evidence$date) %in% month.name[4:5],"apr_may",
                    ifelse(months(Evidence$date) %in% month.name[6:7],"jun_jul",
                    ifelse(months(Evidence$date) %in% month.name[8:9],"aug_sep","blah")))

Evidence$INUN <-    ifelse(q_bin[1]<= Evidence$q_scenario1 & Evidence$q_scenario1<q_bin[2],"bin_1",
                    ifelse(q_bin[2]<= Evidence$q_scenario1 & Evidence$q_scenario1<q_bin[3],"bin_2",
                    ifelse(q_bin[3]<= Evidence$q_scenario1 & Evidence$q_scenario1<q_bin[4],"bin_3",
                    ifelse(q_bin[4]<= Evidence$q_scenario1 & Evidence$q_scenario1<q_bin[5],"bin_4",
                    ifelse(q_bin[5]<= Evidence$q_scenario1 & Evidence$q_scenario1<q_bin[6],"bin_5",
                    ifelse(q_bin[6]<= Evidence$q_scenario1 & Evidence$q_scenario1<q_bin[7],"bin_6",
                    ifelse(q_bin[7]<= Evidence$q_scenario1 & Evidence$q_scenario1<q_bin[8],"bin_7",
                    ifelse(q_bin[8]<= Evidence$q_scenario1 & Evidence$q_scenario1<q_bin[9],"bin_8",
                    ifelse(q_bin[9]<= Evidence$q_scenario1 & Evidence$q_scenario1<q_bin[10],"bin_9",
                    ifelse(q_bin[10]<= Evidence$q_scenario1 & Evidence$q_scenario1<q_bin[11],"bin_10",
                    ifelse(q_bin[11]<= Evidence$q_scenario1 & Evidence$q_scenario1<q_bin[12],"bin_11",
                    ifelse(q_bin[12]<= Evidence$q_scenario1 & Evidence$q_scenario1<q_bin[13],"bin_12",
                    ifelse(q_bin[13]<= Evidence$q_scenario1 & Evidence$q_scenario1<q_bin[14],"bin_13",
                    ifelse(q_bin[14]<= Evidence$q_scenario1 & Evidence$q_scenario1<q_bin[15],"bin_14",
                    ifelse(q_bin[15]<= Evidence$q_scenario1,"bin_15","blah")))))))))))))))


Evidence$RECESSION_RATE <- ifelse(Evidence$rr_scenario1 <= recessionrate_bin[1],"bin_1",
                           ifelse(recessionrate_bin[1]< Evidence$rr_scenario1 & Evidence$rr_scenario1 <=recessionrate_bin[2],"bin_2",
                           ifelse(recessionrate_bin[2]< Evidence$rr_scenario1 & Evidence$rr_scenario1 <=recessionrate_bin[3],"bin_3",
                           ifelse(recessionrate_bin[3]< Evidence$rr_scenario1 & Evidence$rr_scenario1 <=recessionrate_bin[4],"bin_4",
                           ifelse(Evidence$rr_scenario1 > recessionrate_bin[4],"bin_5","blah")))))  

Evidence <- Evidence[-1:-9]
```

####Scenario 2 Evidence####
```{r}
Evidence <- c(hydro[1:nrow(recessionrate_14d),],recessionrate_14d[,2:5])
Evidence <-data.frame(Evidence)
Evidence$TIMING <- NA
Evidence$INUN <- NA
Evidence$RECESSION_RATE<- NA
Evidence <- setNames(Evidence, c("date","q_natural", "q_scenario1","q_scenario2","q_scenario3","rr_natural", "rr_scenario1","rr_scenario2","scenario3", "TIMING","INUN","RECESSION_RATE"))

Evidence$TIMING <-  ifelse(months(Evidence$date) %in% month.name[4:5],"apr_may",
                    ifelse(months(Evidence$date) %in% month.name[6:7],"jun_jul",
                    ifelse(months(Evidence$date) %in% month.name[8:9],"aug_sep","blah")))

Evidence$INUN <-    ifelse(q_bin[1]<= Evidence$q_scenario2 & Evidence$q_scenario2<q_bin[2],"bin_1",
                    ifelse(q_bin[2]<= Evidence$q_scenario2 & Evidence$q_scenario2<q_bin[3],"bin_2",
                    ifelse(q_bin[3]<= Evidence$q_scenario2 & Evidence$q_scenario2<q_bin[4],"bin_3",
                    ifelse(q_bin[4]<= Evidence$q_scenario2 & Evidence$q_scenario2<q_bin[5],"bin_4",
                    ifelse(q_bin[5]<= Evidence$q_scenario2 & Evidence$q_scenario2<q_bin[6],"bin_5",
                    ifelse(q_bin[6]<= Evidence$q_scenario2 & Evidence$q_scenario2<q_bin[7],"bin_6",
                    ifelse(q_bin[7]<= Evidence$q_scenario2 & Evidence$q_scenario2<q_bin[8],"bin_7",
                    ifelse(q_bin[8]<= Evidence$q_scenario2 & Evidence$q_scenario2<q_bin[9],"bin_8",
                    ifelse(q_bin[9]<= Evidence$q_scenario2 & Evidence$q_scenario2<q_bin[10],"bin_9",
                    ifelse(q_bin[10]<= Evidence$q_scenario2 & Evidence$q_scenario2<q_bin[11],"bin_10",
                    ifelse(q_bin[11]<= Evidence$q_scenario2 & Evidence$q_scenario2<q_bin[12],"bin_11",
                    ifelse(q_bin[12]<= Evidence$q_scenario2 & Evidence$q_scenario2<q_bin[13],"bin_12",
                    ifelse(q_bin[13]<= Evidence$q_scenario2 & Evidence$q_scenario2<q_bin[14],"bin_13",
                    ifelse(q_bin[14]<= Evidence$q_scenario2 & Evidence$q_scenario2<q_bin[15],"bin_14",
                    ifelse(q_bin[15]<= Evidence$q_scenario2,"bin_15","blah")))))))))))))))


Evidence$RECESSION_RATE <- ifelse(Evidence$rr_scenario2 <= recessionrate_bin[1],"bin_1",
                           ifelse(recessionrate_bin[1]< Evidence$rr_scenario2 & Evidence$rr_scenario2 <=recessionrate_bin[2],"bin_2",
                           ifelse(recessionrate_bin[2]< Evidence$rr_scenario2 & Evidence$rr_scenario2 <=recessionrate_bin[3],"bin_3",
                           ifelse(recessionrate_bin[3]< Evidence$rr_scenario2 & Evidence$rr_scenario2 <=recessionrate_bin[4],"bin_4",
                           ifelse(Evidence$rr_scenario2 > recessionrate_bin[4],"bin_5","blah")))))  

Evidence <- Evidence[-1:-9]
```

####Scenario 3 Evidence####
```{r}
Evidence <- c(hydro[1:nrow(recessionrate_14d),],recessionrate_14d[,2:5])
Evidence <-data.frame(Evidence)
Evidence$TIMING <- NA
Evidence$INUN <- NA
Evidence$RECESSION_RATE<- NA
Evidence <- setNames(Evidence, c("date","q_natural", "q_scenario1","q_scenario2","q_scenario3","rr_natural", "rr_scenario1","rr_scenario2","rr_scenario3", "TIMING","INUN","RECESSION_RATE"))

Evidence$TIMING <-  ifelse(months(Evidence$date) %in% month.name[4:5],"apr_may",
                    ifelse(months(Evidence$date) %in% month.name[6:7],"jun_jul",
                    ifelse(months(Evidence$date) %in% month.name[8:9],"aug_sep","blah")))

Evidence$INUN <-    ifelse(q_bin[1]<= Evidence$q_scenario3 & Evidence$q_scenario3<q_bin[2],"bin_1",
                    ifelse(q_bin[2]<= Evidence$q_scenario3 & Evidence$q_scenario3<q_bin[3],"bin_2",
                    ifelse(q_bin[3]<= Evidence$q_scenario3 & Evidence$q_scenario3<q_bin[4],"bin_3",
                    ifelse(q_bin[4]<= Evidence$q_scenario3 & Evidence$q_scenario3<q_bin[5],"bin_4",
                    ifelse(q_bin[5]<= Evidence$q_scenario3 & Evidence$q_scenario3<q_bin[6],"bin_5",
                    ifelse(q_bin[6]<= Evidence$q_scenario3 & Evidence$q_scenario3<q_bin[7],"bin_6",
                    ifelse(q_bin[7]<= Evidence$q_scenario3 & Evidence$q_scenario3<q_bin[8],"bin_7",
                    ifelse(q_bin[8]<= Evidence$q_scenario3 & Evidence$q_scenario3<q_bin[9],"bin_8",
                    ifelse(q_bin[9]<= Evidence$q_scenario3 & Evidence$q_scenario3<q_bin[10],"bin_9",
                    ifelse(q_bin[10]<= Evidence$q_scenario3 & Evidence$q_scenario3<q_bin[11],"bin_10",
                    ifelse(q_bin[11]<= Evidence$q_scenario3 & Evidence$q_scenario3<q_bin[12],"bin_11",
                    ifelse(q_bin[12]<= Evidence$q_scenario3 & Evidence$q_scenario3<q_bin[13],"bin_12",
                    ifelse(q_bin[13]<= Evidence$q_scenario3 & Evidence$q_scenario3<q_bin[14],"bin_13",
                    ifelse(q_bin[14]<= Evidence$q_scenario3 & Evidence$q_scenario3<q_bin[15],"bin_14",
                    ifelse(q_bin[15]<= Evidence$q_scenario3,"bin_15","blah")))))))))))))))


Evidence$RECESSION_RATE <- ifelse(Evidence$rr_scenario3 <= recessionrate_bin[1],"bin_1",
                           ifelse(recessionrate_bin[1]< Evidence$rr_scenario3 & Evidence$rr_scenario3 <=recessionrate_bin[2],"bin_2",
                           ifelse(recessionrate_bin[2]< Evidence$rr_scenario3 & Evidence$rr_scenario3 <=recessionrate_bin[3],"bin_3",
                           ifelse(recessionrate_bin[3]< Evidence$rr_scenario3 & Evidence$rr_scenario3 <=recessionrate_bin[4],"bin_4",
                           ifelse(Evidence$rr_scenario3 > recessionrate_bin[4],"bin_5","blah")))))  

Evidence <- Evidence[-1:-9]
```

Binning evidence based on flow bins and trimming data not within the discrete timing states. 
```{r}
Evidence_q1 <- Evidence[Evidence[2] =="bin_1" & Evidence[1]!="blah",]
Evidence_q2 <- Evidence[Evidence[2] =="bin_2" & Evidence[1]!="blah",]
Evidence_q3 <- Evidence[Evidence[2] =="bin_3" & Evidence[1]!="blah",]
Evidence_q4 <- Evidence[Evidence[2] =="bin_4" & Evidence[1]!="blah",]
Evidence_q5 <- Evidence[Evidence[2] =="bin_5" & Evidence[1]!="blah",]
Evidence_q6 <- Evidence[Evidence[2] =="bin_6" & Evidence[1]!="blah",]
Evidence_q7 <- Evidence[Evidence[2] =="bin_7" & Evidence[1]!="blah",]
Evidence_q8 <- Evidence[Evidence[2] =="bin_8" & Evidence[1]!="blah",]
Evidence_q9 <- Evidence[Evidence[2] =="bin_9" & Evidence[1]!="blah",]
Evidence_q10 <- Evidence[Evidence[2] =="bin_10" & Evidence[1]!="blah",]
Evidence_q11 <- Evidence[Evidence[2] =="bin_11" & Evidence[1]!="blah",]
Evidence_q12 <- Evidence[Evidence[2] =="bin_12" & Evidence[1]!="blah",]
Evidence_q13 <- Evidence[Evidence[2] =="bin_13" & Evidence[1]!="blah",]
Evidence_q14 <- Evidence[Evidence[2] =="bin_14" & Evidence[1]!="blah",]
Evidence_q15 <- Evidence[Evidence[2] =="bin_15" & Evidence[1]!="blah",]
```

### Calculate posterior recruitment probability given that an event withing the specified q_bin occurs by instantiating the BN with evidence ###
```{r}
Recruit_q1<-data.frame()
for (i in 1:(nrow(Evidence_q1))) 
{Recruit_q1 <-rbind(Recruit_q1, cpquery(bnfit,event = (RECRUIT =="YES"),
                                        evidence = list(TIMING=Evidence_q1[i,1],
                                                        INUN=Evidence_q1[i,2],
                                                        RECESSION_RATE = Evidence_q1[i,3]), 
                                                        method = "lw",
                                                        n=25000 ))}

Recruit_q2<-data.frame()
for (i in 1:(nrow(Evidence_q2))) 
{Recruit_q2 <-rbind(Recruit_q2, cpquery(bnfit,event = (RECRUIT =="YES"),
                                        evidence = list(TIMING=Evidence_q2[i,1],
                                                        INUN=Evidence_q2[i,2],
                                                        RECESSION_RATE = Evidence_q2[i,3]), 
                                                        method = "lw",
                                                        n=25000 ))}

Recruit_q3<-data.frame()
for (i in 1:(nrow(Evidence_q3))) 
{Recruit_q3 <-rbind(Recruit_q3, cpquery(bnfit,event = (RECRUIT =="YES"),
                                        evidence = list(TIMING=Evidence_q3[i,1],
                                                        INUN=Evidence_q3[i,2],
                                                        RECESSION_RATE = Evidence_q3[i,3]), 
                                                        method = "lw",
                                                        n=25000 ))}

Recruit_q4<-data.frame()
for (i in 1:(nrow(Evidence_q4))) 
{Recruit_q4 <-rbind(Recruit_q4, cpquery(bnfit,event = (RECRUIT =="YES"),
                                        evidence = list(TIMING=Evidence_q4[i,1],
                                                        INUN=Evidence_q4[i,2],
                                                        RECESSION_RATE = Evidence_q4[i,3]), 
                                                        method = "lw",
                                                        n=2500))}

Recruit_q5<-data.frame()
for (i in 1:(nrow(Evidence_q5))) 
{Recruit_q5 <-rbind(Recruit_q5, cpquery(bnfit,event = (RECRUIT =="YES"),
                                        evidence = list(TIMING=Evidence_q5[i,1],
                                                        INUN=Evidence_q5[i,2],
                                                        RECESSION_RATE = Evidence_q5[i,3]), 
                                                        method = "lw",
                                                        n=25000 ))}

Recruit_q6<-data.frame()
for (i in 1:(nrow(Evidence_q6))) 
{Recruit_q6 <-rbind(Recruit_q6, cpquery(bnfit,event = (RECRUIT =="YES"),
                                        evidence = list(TIMING=Evidence_q6[i,1],
                                                        INUN=Evidence_q6[i,2],
                                                        RECESSION_RATE = Evidence_q6[i,3]), 
                                                        method = "lw",
                                                        n=25000 ))}

Recruit_q7<-data.frame()
for (i in 1:(nrow(Evidence_q7))) 
{Recruit_q7 <-rbind(Recruit_q7, cpquery(bnfit,event = (RECRUIT =="YES"),
                                        evidence = list(TIMING=Evidence_q7[i,1],
                                                        INUN=Evidence_q7[i,2],
                                                        RECESSION_RATE = Evidence_q7[i,3]), 
                                                        method = "lw",
                                                        n=25000 ))}

Recruit_q8<-data.frame()
for (i in 1:(nrow(Evidence_q8))) 
{Recruit_q8 <-rbind(Recruit_q8, cpquery(bnfit,event = (RECRUIT =="YES"),
                                        evidence = list(TIMING=Evidence_q8[i,1],
                                                        INUN=Evidence_q8[i,2],
                                                        RECESSION_RATE = Evidence_q8[i,3]), 
                                                        method = "lw",
                                                        n=25000 ))}

Recruit_q9<-data.frame()
for (i in 1:(nrow(Evidence_q9))) 
{Recruit_q9 <-rbind(Recruit_q9, cpquery(bnfit,event = (RECRUIT =="YES"),
                                        evidence = list(TIMING=Evidence_q9[i,1],
                                                        INUN=Evidence_q9[i,2],
                                                        RECESSION_RATE = Evidence_q9[i,3]), 
                                                        method = "lw",
                                                        n=25000 ))}

Recruit_q10<-data.frame()
for (i in 1:(nrow(Evidence_q10))) 
{Recruit_q10 <-rbind(Recruit_q10, cpquery(bnfit,event = (RECRUIT =="YES"),
                                        evidence = list(TIMING=Evidence_q10[i,1],
                                                        INUN=Evidence_q10[i,2],
                                                        RECESSION_RATE = Evidence_q10[i,3]), 
                                                        method = "lw",
                                                        n=25000 ))}

Recruit_q11<-data.frame()
for (i in 1:(nrow(Evidence_q11))) 
{Recruit_q11 <-rbind(Recruit_q11, cpquery(bnfit,event = (RECRUIT =="YES"),
                                        evidence = list(TIMING=Evidence_q11[i,1],
                                                        INUN=Evidence_q11[i,2],
                                                        RECESSION_RATE = Evidence_q11[i,3]), 
                                                        method = "lw",
                                                        n=25000 ))}

Recruit_q12<-data.frame()
for (i in 1:(nrow(Evidence_q12))) 
{Recruit_q12 <-rbind(Recruit_q12, cpquery(bnfit,event = (RECRUIT =="YES"),
                                        evidence = list(TIMING=Evidence_q12[i,1],
                                                        INUN=Evidence_q12[i,2],
                                                        RECESSION_RATE = Evidence_q12[i,3]), 
                                                        method = "lw",
                                                        n=25000 ))}

Recruit_q13<-data.frame()
for (i in 1:(nrow(Evidence_q13))) 
{Recruit_q13 <-rbind(Recruit_q13, cpquery(bnfit,event = (RECRUIT =="YES"),
                                        evidence = list(TIMING=Evidence_q13[i,1],
                                                        INUN=Evidence_q13[i,2],
                                                        RECESSION_RATE = Evidence_q13[i,3]), 
                                                        method = "lw",
                                                        n=25000 ))}

Recruit_q14<-data.frame()
for (i in 1:(nrow(Evidence_q14))) 
{Recruit_q14 <-rbind(Recruit_q14, cpquery(bnfit,event = (RECRUIT =="YES"),
                                        evidence = list(TIMING=Evidence_q14[i,1],
                                                        INUN=Evidence_q14[i,2],
                                                        RECESSION_RATE = Evidence_q14[i,3]), 
                                                        method = "lw",
                                                        n=25000 ))}

Recruit_q15<-data.frame()
for (i in 1:(nrow(Evidence_q15))) 
{Recruit_q15 <-rbind(Recruit_q15, cpquery(bnfit,event = (RECRUIT =="YES"),
                                        evidence = list(TIMING=Evidence_q15[i,1],
                                                        INUN=Evidence_q15[i,2],
                                                        RECESSION_RATE = Evidence_q15[i,3]), 
                                                        method = "lw",
                                                        n=25000 ))}

```

Two output files,(1) The recruitment probability for each q_bin and (2) the recruitment probability for each cell, are generated, written, and saved in the output folder. Result/output files can be re-written to specify which scenario is being written to .csv for clarity (e.g. q_cells_natural_recruit.csv). The file path can also be changed as needed.

The code chunk below created dataframe Recruit_prob which summarizes the recruitment probability for each q_bin and adds a recruitment probability column to the q_cells dataframe.
```{r}

recruit_prob<- c(mean(Recruit_q1[,1]),mean(Recruit_q2[,1]),mean(Recruit_q3[,1]),mean(Recruit_q4[,1]),mean(Recruit_q5[,1]),mean(Recruit_q6[,1]),mean(Recruit_q7[,1]),
                 mean(Recruit_q8[,1]),mean(Recruit_q9[,1]),mean(Recruit_q10[,1]),mean(Recruit_q11[,1]),mean(Recruit_q12[,1]),mean(Recruit_q13[,1]),mean(Recruit_q14[,1]),
                 mean(Recruit_q15[,1]))
recruit_prob <-data.frame(recruit_prob)
dimnames(recruit_prob) = list("bin"=c("bin_1", "bin_2", "bin_3", "bin_4", "bin_5","bin_6","bin_7",
                                      "bin_8", "bin_9","bin_10","bin_11","bin_12","bin_13","bin_14","bin_15"),"P(R|q)")

q_cells$recruit_q <-    ifelse(q_cells$bin =="bin_1",recruit_prob[1,],
                    ifelse(q_cells$bin =="bin_2",recruit_prob[2,],
                    ifelse(q_cells$bin =="bin_3",recruit_prob[3,],
                    ifelse(q_cells$bin =="bin_4",recruit_prob[4,],
                    ifelse(q_cells$bin =="bin_5",recruit_prob[5,],
                    ifelse(q_cells$bin =="bin_6",recruit_prob[6,],
                    ifelse(q_cells$bin =="bin_7",recruit_prob[7,],
                    ifelse(q_cells$bin =="bin_8",recruit_prob[8,],
                    ifelse(q_cells$bin =="bin_9",recruit_prob[9,],
                    ifelse(q_cells$bin =="bin_10",recruit_prob[10,],
                    ifelse(q_cells$bin =="bin_11",recruit_prob[11,],
                    ifelse(q_cells$bin =="bin_12",recruit_prob[12,],
                    ifelse(q_cells$bin =="bin_13",recruit_prob[13,],
                    ifelse(q_cells$bin =="bin_14",recruit_prob[14,],
                    ifelse(q_cells$bin =="bin_15",recruit_prob[15,],
                    "blah")))))))))))))))
```

#Natural#
```{r}

recruit_natural<-data.frame()
for (i in 1:length(q_prob_natural_yr))
{recruit_natural <-rbind(recruit_natural, sum(q_prob_yr[1,i:15]*recruit_prob[i:15,1]))
print(recruit_natural)}
recruit_natural<-cbind(recruit_natural, recruit_prob)
rownames(recruit_natural) <- c("bin_1", "bin_2", "bin_3", "bin_4", "bin_5","bin_6","bin_7",
                               "bin_8", "bin_9","bin_10","bin_11","bin_12","bin_13","bin_14","bin_15")
colnames(recruit_natural) <- c("P(R)", "P(R|q)")

q_cells$recruit <-  ifelse(q_cells$bin =="bin_1",recruit_natural[1,1],
                    ifelse(q_cells$bin =="bin_2",recruit_natural[2,1],
                    ifelse(q_cells$bin =="bin_3",recruit_natural[3,1],
                    ifelse(q_cells$bin =="bin_4",recruit_natural[4,1],
                    ifelse(q_cells$bin =="bin_5",recruit_natural[5,1],
                    ifelse(q_cells$bin =="bin_6",recruit_natural[6,1],
                    ifelse(q_cells$bin =="bin_7",recruit_natural[7,1],
                    ifelse(q_cells$bin =="bin_8",recruit_natural[8,1],
                    ifelse(q_cells$bin =="bin_9",recruit_natural[9,1],
                    ifelse(q_cells$bin =="bin_10",recruit_natural[10,1],
                    ifelse(q_cells$bin =="bin_11",recruit_natural[11,1],
                    ifelse(q_cells$bin =="bin_12",recruit_natural[12,1],
                    ifelse(q_cells$bin =="bin_13",recruit_natural[13,1],
                    ifelse(q_cells$bin =="bin_14",recruit_natural[14,1],
                           "blah"))))))))))))))
```


#Scenario1#
```{r}

recruit_scenario1<-data.frame()
for (i in 1:length(q_prob_scenario1_yr))
{recruit_scenario1 <-rbind(recruit_scenario1, sum(q_prob_scenario1_yr[,i:15]*recruit_prob[i:15,1]))
print(recruit_scenario1)}
recruit_scenario1<-cbind(recruit_scenario1, recruit_prob)
rownames(recruit_scenario1) <- c("bin_1", "bin_2", "bin_3", "bin_4", "bin_5","bin_6","bin_7",
                               "bin_8", "bin_9","bin_10","bin_11","bin_12","bin_13","bin_14","bin_15")
colnames(recruit_scenario1) <- c("P(R)", "P(R|q)")

q_cells$recruit <-  ifelse(q_cells$bin =="bin_1",recruit_scenario1[1,1],
                    ifelse(q_cells$bin =="bin_2",recruit_scenario1[2,1],
                    ifelse(q_cells$bin =="bin_3",recruit_scenario1[3,1],
                    ifelse(q_cells$bin =="bin_4",recruit_scenario1[4,1],
                    ifelse(q_cells$bin =="bin_5",recruit_scenario1[5,1],
                    ifelse(q_cells$bin =="bin_6",recruit_scenario1[6,1],
                    ifelse(q_cells$bin =="bin_7",recruit_scenario1[7,1],
                    ifelse(q_cells$bin =="bin_8",recruit_scenario1[8,1],
                    ifelse(q_cells$bin =="bin_9",recruit_scenario1[9,1],
                    ifelse(q_cells$bin =="bin_10",recruit_scenario1[10,1],
                    ifelse(q_cells$bin =="bin_11",recruit_scenario1[11,1],
                    ifelse(q_cells$bin =="bin_12",recruit_scenario1[12,1],
                    ifelse(q_cells$bin =="bin_13",recruit_scenario1[13,1],
                    ifelse(q_cells$bin =="bin_14",recruit_scenario1[14,1],
                    ifelse(q_cells$bin =="bin_15",recruit_scenario1[15,1],
                           "blah")))))))))))))))
```

#Scenario2#
```{r}

recruit_scenario2<-data.frame()
for (i in 1:length(q_prob_natural_yr))
{recruit_scenario2 <-rbind(recruit_scenario2, sum(q_prob_natural_yr[,i:15]*recruit_prob[i:15,1]))
print(recruit_scenario2)}
recruit_scenario2<-cbind(recruit_scenario2, recruit_prob)
rownames(recruit_scenario2) <- c("bin_1", "bin_2", "bin_3", "bin_4", "bin_5","bin_6","bin_7",
                               "bin_8", "bin_9","bin_10","bin_11","bin_12","bin_13","bin_14","bin_15")
colnames(recruit_scenario2) <- c("P(R)", "P(R|q)")

q_cells$recruit <-  ifelse(q_cells$bin =="bin_1",recruit_scenario2[1,1],
                    ifelse(q_cells$bin =="bin_2",recruit_scenario2[2,1],
                    ifelse(q_cells$bin =="bin_3",recruit_scenario2[3,1],
                    ifelse(q_cells$bin =="bin_4",recruit_scenario2[4,1],
                    ifelse(q_cells$bin =="bin_5",recruit_scenario2[5,1],
                    ifelse(q_cells$bin =="bin_6",recruit_scenario2[6,1],
                    ifelse(q_cells$bin =="bin_7",recruit_scenario2[7,1],
                    ifelse(q_cells$bin =="bin_8",recruit_scenario2[8,1],
                    ifelse(q_cells$bin =="bin_9",recruit_scenario2[9,1],
                    ifelse(q_cells$bin =="bin_10",recruit_scenario2[10,1],
                    ifelse(q_cells$bin =="bin_11",recruit_scenario2[11,1],
                    ifelse(q_cells$bin =="bin_12",recruit_scenario2[12,1],
                    ifelse(q_cells$bin =="bin_13",recruit_scenario2[13,1],
                    ifelse(q_cells$bin =="bin_14",recruit_scenario2[14,1],
                    ifelse(q_cells$bin =="bin_15",recruit_scenario2[15,1],
                           "blah")))))))))))))))
```

#Scenario3#
```{r}

recruit_scenario3<-data.frame()
for (i in 1:length(q_prob_natural_yr))
{recruit_scenario3 <-rbind(recruit_scenario3, sum(q_prob_natural_yr[,i:15]*recruit_prob[i:15,1]))
print(recruit_scenario3)}
recruit_scenario3<-cbind(recruit_scenario3, recruit_prob)
rownames(recruit_scenario3) <- c("bin_1", "bin_2", "bin_3", "bin_4", "bin_5","bin_6","bin_7",
                               "bin_8", "bin_9","bin_10","bin_11","bin_12","bin_13","bin_14","bin_15")
colnames(recruit_scenario3) <- c("P(R)", "P(R|q)")

q_cells$recruit <-  ifelse(q_cells$bin =="bin_1",recruit_scenario3[1,1],
                    ifelse(q_cells$bin =="bin_2",recruit_scenario3[2,1],
                    ifelse(q_cells$bin =="bin_3",recruit_scenario3[3,1],
                    ifelse(q_cells$bin =="bin_4",recruit_scenario3[4,1],
                    ifelse(q_cells$bin =="bin_5",recruit_scenario3[5,1],
                    ifelse(q_cells$bin =="bin_6",recruit_scenario3[6,1],
                    ifelse(q_cells$bin =="bin_7",recruit_scenario3[7,1],
                    ifelse(q_cells$bin =="bin_8",recruit_scenario3[8,1],
                    ifelse(q_cells$bin =="bin_9",recruit_scenario3[9,1],
                    ifelse(q_cells$bin =="bin_10",recruit_scenario3[10,1],
                    ifelse(q_cells$bin =="bin_11",recruit_scenario3[11,1],
                    ifelse(q_cells$bin =="bin_12",recruit_scenario3[12,1],
                    ifelse(q_cells$bin =="bin_13",recruit_scenario3[13,1],
                    ifelse(q_cells$bin =="bin_14",recruit_scenario3[14,1],
                    ifelse(q_cells$bin =="bin_15",recruit_scenario3[15,1],
                           "blah")))))))))))))))
```
Two output files,(1) The recruitment probability for each q_bin and (2) the recruitment probability for each cell, are generated, written, and saved in the output folder. Result/output files can be re-written to specify which scenario is being written to .csv for clarity (e.g. q_cells_natural_recruit.csv). The file path can also be changed as needed.


The code chunck below writes output files for the natural scenario.
```{r}
write.csv(recruit_natural,file = "output/natural_s2_recruit_prob_TEST.csv")
write.csv(q_cells, file = "output/natural_s2_q_cells_recruit_TEST.csv",row.names= FALSE)
```


The code chunck below writes output files for scenario 1.
```{r}
write.csv(recruit_scenario1,file = "output/scenario1_s2_recruit_prob.csv")
write.csv(q_cells,file = "output/scenario1_s2_q_cells_recruit.csv",row.names= FALSE)
```


The code chunck below writes output files for scenario 2.
```{r}
write.csv(recruit_scenario2,file = "output/scenario2_s2_recruit_prob.csv")
write.csv(q_cells,file = "output/scenario2_s2_q_cells_recruit.csv",row.names= FALSE)
```


The code chunck below writes output files for scenario 3.
```{r}
write.csv(recruit_scenario3,file = "output/scenario3_s2_recruit_prob.csv")
write.csv(q_cells,file = "output/scenario3_s2_q_cells_recruit.csv",row.names= FALSE)
```


###References ###
Gentleman R, Whalen E, Huber W and Falcon S (2017). graph: graph: A package to handle graph data structures. R package version 1.54.0.

Hansen KD, Gentry J, Long L, Gentleman R, Falcon S, Hahne F and Sarkar D (2017). Rgraphviz: Provides plotting capabilities for R graph objects. R package version 2.20.0.

Huber, W., Carey, J. V, Gentleman, R., Anders, S., Carlson, M., Carvalho, S. B, Bravo, C. H, Davis, S., Gatto, L., Girke, T., Gottardo, R., Hahne, F., Hansen, D. K, Irizarry, A. R, Lawrence, M., Love, I. M, MacDonald, J., Obenchain, V., Ole's, K. A, Pag'es, H., Reyes, A., Shannon, P., Smyth, K. G, Tenenbaum, D., Waldron, L., Morgan and M. (2015). “Orchestrating high-throughput genomic analysis with Bioconductor.” Nature Methods, 12(2), pp. 115–121. http://www.nature.com/nmeth/journal/v12/n2/full/nmeth.3252.html.

Koller D, Friedman N (2009). Probabilistic Graphical Models: Principles and Techniques. MIT Press.

Pearl J (2009). Causality: Models, Reasoning and Inference. Cambridge University Press, 2nd edition.

Nagarajan, Radhakrishnan, Marco Scutari, and Sophie Lèbre. Bayesian Networks in R. New York, NY: Springer New York, 2013. http://link.springer.com/10.1007/978-1-4614-6446-4. 
http://www.bnlearn.com

Scutari, Marco (2018). bnlearn: Bayesian Network Structure Learning, Parameter Learning and Inference. R package version 4.4 https://cran.r-project.org/web/packages/bnlearn/

https://stackoverflow.com/questions/40364458/r-some-conditional-probability-distributions-of-node-do-not-sum-to-one

example - https://stackoverflow.com/questions/34127885/bnlearn-wrong-dimensions-for-node

cpquery - http://www.bnlearn.com/documentation/man/cpquery.html

soft/hard evidence - https://pdfs.semanticscholar.org/13f5/795c9d4075d4468131fe9405bae5ec3e3857.pdf
