==> 4-gila-analysis_new.Rmd <==
---
title: Upper Gila Watershed Bayesian Network Results Analysis
author: Lauren Victoria Jaramillo, UNM (alwayslo@unm.edu)
date: 4/5/2019
update: 8/20/2019
output: html_notebook
---

This R notebook is used to used to analyze the results from the BN model.

###Result Analysis###
```{r}
rm(list=ls())

library("bnlearn")
library("readr")
library("Rgraphviz")
library("xts")
library("prob")
library("TTR")
library("bigmemory")
library("zoo")
library("raster")
library("sp")
library("rgdal")
library("rasterVis")
library("ggplot2")
library("rgdal")
# library("dplyr")
library("methods")
require("BiocGenerics")
require("graph")
library("sf")
library("maps")
library("RColorBrewer")
library("stringi", lib.loc="/Library/Frameworks/R.framework/Versions/3.4/Resources/library")
getwd()
```

Load and plot all flow scenarios modeled.
```{r}
hydro <- read.csv("data/scenarios_revised.csv", skip=1,header = TRUE)[ ,1:5]
hydro <- setNames(hydro, c("date","natural", "Alt. E","Alt. B","Alt. A"))
hydro$date <-as.Date(hydro$date, format="%m/%d/%Y")
hydro$julian <- format(hydro$date, "%j")
hydro$julian <-stri_replace_all_regex(hydro$julian, "\\b0*(\\d+)\\b", "$1")
hydro$month <- format(hydro$date,"%m")
```

Plots the hydrologic data that is actually used in the model which is governed by the lowest minimum innundation value.

Plots the mean daily flows for all four scenarios (natural,current_operations,combined_operations,combined_operations_AWSA).
```{r}
meanJulian <- data.frame()

for (i in 1:length(unique(hydro$julian))) 
{ julian_i <- colMeans(hydro[hydro$julian== i,2:5])
  meanJulian <- rbind(meanJulian, julian_i)}

meanJulian <- cbind( 1:length(unique(hydro$julian)),meanJulian)
meanJulian<-setNames(meanJulian, c("julian","natural", "Alt_E","Alt_B","Alt_A"))
meanJulian<-na.omit(meanJulian)

ggplot(meanJulian, aes(x=julian))+ 
  geom_line(aes(y = natural, color = "natural"),linetype = 1)+
  geom_line(aes(y = Alt_E, color = "blue"),linetype = 1)+
  geom_line(aes(y = Alt_B, color = "green"),linetype = 2)+
  geom_line(aes(y = Alt_A, color = "orange"),linetype = 4)+
  xlab("Julian day")+
  ylab("flow (cfs)")+
  ggtitle("Upper Gila Mean Daily Flow\n April to September 1936 - 2017")+
  theme(plot.title = element_text(hjust = 0.5))+
  labs(color="scenario")+
  geom_hline(yintercept = 1000)

annual_vol <- data.frame(colSums(meanJulian[2:5]*8640/43560))
annual_vol <-setNames(annual_vol,("acre-ft/yr"))
```




Plots the one standard deviation above the mean daily flows for all four scenarios (natural,current_operations,combined_operations,combined_operations_AWSA).
```{r}
sigma1Julian <- data.frame()

for (i in 1:length(unique(hydro$julian))) 
{ julian_i <- colMeans(hydro[hydro$julian== i,2:5])+ colStdevs(hydro[hydro$julian== i,2:5])
  sigma1Julian <- rbind(sigma1Julian, julian_i)}

sigma1Julian <- cbind( 1:length(unique(hydro$julian)),sigma1Julian)
sigma1Julian<-setNames(sigma1Julian, c("julian","natural", "Alt_E","Alt_B","Alt_A"))
sigma1Julian<-na.omit(sigma1Julian)

ggplot(sigma1Julian, aes(x=julian))+ 
  geom_line(aes(y = natural, color = "natural"),linetype = 1)+
  geom_line(aes(y = Alt_E, color = "Alt. E"),linetype = 1)+
  geom_line(aes(y = Alt_B, color = "Alt. B"),linetype = 2)+
  geom_line(aes(y = Alt_A, color = "Alt. A"),linetype = 4)+
  xlab("Julian day")+
  ylab("flow (cfs)")+
  ggtitle("Upper Gila Standard Deviation Above Mean Daily Flow\n April to September 1936 - 2017")+
  theme(plot.title = element_text(hjust = 0.5))+
  labs(color="scenario")+
  geom_hline(yintercept = c(1000, 1500, 2000, 2500, 3000, 3500, 4000))

ggsave("meandailyflow.tiff", units="in", width=5, height=4, dpi=900, compression = 'lzw')


```

Plots the one standard deviation above the mean daily flows for all four scenarios (natural,current_operations,combined_operations,combined_operations_AWSA).
```{r}
sigma2Julian <- data.frame()

for (i in 1:length(unique(hydro$julian))) 
{ julian_i <- colMeans(hydro[hydro$julian== i,2:5])+ 2*colStdevs(hydro[hydro$julian== i,2:5])
  sigma2Julian <- rbind(sigma2Julian, julian_i)}

sigma2Julian <- cbind( 1:length(unique(hydro$julian)),sigma2Julian)
sigma2Julian<-setNames(sigma2Julian, c("julian","natural", "Alt_E","Alt_B","Alt_A"))
sigma2Julian<-na.omit(sigma2Julian)

ggplot(sigma2Julian, aes(x=julian))+ 
  geom_line(aes(y = natural, color = "natural"),linetype = 1)+
  geom_line(aes(y = Alt_E, color = "Alt. E"),linetype = 1)+
  geom_line(aes(y = Alt_B, color = "Alt. B"),linetype = 2)+
  geom_line(aes(y = Alt_A, color = "Alt. A"),linetype = 4)+
  xlab("Julian day")+
  ylab("flow (cfs)")+
  ggtitle("Upper Gila Two Standard Deviations Above Mean Daily Flow\n April to September 1936 - 2017")+
  theme(plot.title = element_text(hjust = 0.5))+
  labs(color="scenario")+
  geom_hline(yintercept = c(1000, 1500, 2000, 2500, 3000, 3500, 4000))

```


Plot only flows greater than 1000 cfs. Visualize data being used. (not working)
```{r}
naturalHI <- hydro[hydro[2] >= 1000,]
scenario1HI <- hydro[hydro[3] >= 1000,]
scenario2HI <- hydro[hydro[4] >= 1000,]
scenario3HI <- hydro[hydro[5] >= 1000,]

mean_qHI <- data.frame()

for (i in 1:length(unique(scenario3HI$julian))) 
{ julian_i <- colMeans(scenario3HI[scenario3HI$julian== i,2:5])
  mean_qHI <- rbind(mean_qHI, julian_i)}

mean_qHI <- cbind( 1:length(unique(scenario3HI$julian)),mean_qHI)
mean_qHI<-setNames(mean_qHI, c("julian","natural", "Alt_E","Alt_B","Alt_A"))
mean_qHI<-na.omit(mean_qHI)


ggplot(mean_naturalHI, aes(x=julian))+ 
  geom_line(aes(y = natural, color = "natural"),linetype = 1)+
  geom_line(aes(y = Alt_E, color = "Alt. E"),linetype = 1)+
  geom_line(aes(y = Alt_B, color = "Alt. B"),linetype = 2)+
  geom_line(aes(y = Alt_A, color = "Alt. A"),linetype = 4)+
  xlab("Julian day")+
  ylab("flow (cfs)")+
  ggtitle("Upper Gila Mean Daily Flow\n April to September 1936 - 2017")+
  theme(plot.title = element_text(hjust = 0.5))+
  labs(color="scenario")+
  geom_hline(yintercept = 1000)
# 
#
```

Plots mean daily divert flow by scenario and calculates the total volume of diverted flow in acre-ft/year. (not working)
```{r}
q_diff <- cbind(hydro$julian, hydro[2]-hydro[3],hydro[2]-hydro[4],hydro[2]-hydro[5])
q_diff <- setNames(q_diff, c("julian", "Alt. E","Alt. B","Alt. A"))

meanJulianDiff <- data.frame()

for (i in 1:length(unique(q_diff$julian))) 
{ julian_i <- colMeans(q_diff[q_diff$julian== i,2:4])
  meanJulianDiff <- rbind(meanJulianDiff, julian_i)}

meanJulianDiff <- cbind( 1:length(unique(q_diff$julian)),meanJulianDiff)
meanJulianDiff<-setNames(meanJulianDiff, c("julian", "Alt. E","Alt. B","Alt. A"))
meanJulianDiff<-na.omit(meanJulianDiff)

ggplot(meanJulianDiff, aes(x=julian))+ 
  geom_line(aes(y = natural, color = "natural"),linetype = 1)+
  geom_line(aes(y = Alt_E, color = "Alt. E"),linetype = 1)+
  geom_line(aes(y = Alt_B, color = "Alt. B"),linetype = 2)+
  geom_line(aes(y = Alt_A, color = "Alt. A"),linetype = 4)+
  geom_vline(xintercept = 92, linetype="dotted",  color = "black", size=0.5)+
  geom_vline(xintercept = 274, linetype="dotted",  color = "black", size=0.5)+
  xlab("Julian day")+
  ylab("diverted flow (cfs)")+
  ggtitle("Upper Gila Mean Daily Diverted Flow\n April to September 1936 - 2017")+
  theme(plot.title = element_text(hjust = 0.5))+
  labs(color="scenario")

scenario1_divert <- colSums(meanJulianDiff[2]*8640/43560)
scenario2_divert <- colSums(meanJulianDiff[3]*8640/43560)
scenario3_divert <- colSums(meanJulianDiff[4]*8640/43560)

```



<!-- Load scenario results for site 1 (North) to be used for analysis. -->
<!-- ```{r} -->
<!-- r_cells_natural_s1 <- read.csv(file ="/Users/lohebert/GilaBN/output/natural_s1_q_cells_recruit.csv", header=TRUE, sep=",",  -->
<!--                                col.names =  c("cell","flow","bin","recruit_q","recruit")) -->
<!-- r_cells_scenario1_s1 <-read.csv(file ="/Users/lohebert/GilaBN/output/scenario1_s1_q_cells_recruit.csv", header=TRUE, sep=",",  -->
<!--                                 col.names = c("cell","flow","bin","recruit_q","recruit")) -->
<!-- r_cells_scenario2_s1 <- read.csv(file ="/Users/lohebert/GilaBN/output/scenario2_s1_q_cells_recruit.csv", header=TRUE, sep=",",  -->
<!--                                  col.names = c("cell","flow","bin","recruit_q","recruit")) -->
<!-- r_cells_scenario3_s1 <- read.csv(file ="/Users/lohebert/GilaBN/output/scenario3_s1_q_cells_recruit.csv", header=TRUE, sep=",", -->
<!--                                  col.names = c("cell","flow","bin","recruit_q","recruit")) -->

<!-- r_cells_s1 <- data.frame(cbind(r_cells_natural_s1[,1],r_cells_natural_s1[,5],r_cells_scenario1_s1[,5],r_cells_scenario2_s1[,5], -->
<!--                                r_cells_scenario3_s1[,5])) -->
<!-- colnames(r_cells_s1) <- c("cell","natural","Alt. E","Alt. B","Alt. A") -->

<!-- r_prob_natural_s1 <- read.csv(file ="/Users/lohebert/GilaBN/output/natural_s1_recruit_prob.csv", header=TRUE, sep=",") -->
<!-- r_prob_scenario1_s1 <-read.csv(file ="/Users/lohebert/GilaBN/output/scenario1_s1_recruit_prob.csv", header=TRUE, sep=",") -->
<!-- r_prob_scenario2_s1 <- read.csv(file ="/Users/lohebert/GilaBN/output/scenario2_s1_recruit_prob.csv", header=TRUE, sep=",") -->
<!-- r_prob_scenario3_s1 <- read.csv(file ="/Users/lohebert/GilaBN/output/scenario3_s1_recruit_prob.csv", header=TRUE, sep=",") -->

<!-- r_prob_s1 <- cbind(r_prob_natural_s1[,2], r_prob_scenario1_s1[,2],r_prob_scenario2_s1[,2],r_prob_scenario3_s1[,2]) -->
<!-- colnames(r_prob_s1)<- c("natural","Alt. E","Alt. B","Alt. A")   -->
<!-- rownames(r_prob_s1)<- c("bin_1", "bin_2", "bin_3", "bin_4", "bin_5","bin_6","bin_7","bin_8","bin_9", "bin_10", "bin_11", "bin_12", "bin_13","bin_14","bin_15") -->

<!-- ``` -->


Load scenario results for site 2 (Middle) to be used for analysis.
```{r}
r_cells_natural_s2 <- read.csv(file ="/Users/lohebert/GilaBN/output/natural_s2_q_cells_recruit.csv", header=TRUE, sep=",", 
                               col.names =  c("cell","flow","bin","recruit_q","recruit"))
r_cells_scenario1_s2 <-read.csv(file ="/Users/lohebert/GilaBN/output/scenario1_s2_q_cells_recruit.csv", header=TRUE, sep=",", 
                                col.names = c("cell","flow","bin","recruit_q","recruit"))
r_cells_scenario2_s2 <- read.csv(file ="/Users/lohebert/GilaBN/output/scenario2_s2_q_cells_recruit.csv", header=TRUE, sep=",", 
                                 col.names = c("cell","flow","bin","recruit_q","recruit"))
r_cells_scenario3_s2 <- read.csv(file ="/Users/lohebert/GilaBN/output/scenario3_s2_q_cells_recruit.csv", header=TRUE, sep=",", 
                                 col.names = c("cell","flow","bin","recruit_q","recruit"))

r_cells_s2 <- data.frame(cbind(r_cells_natural_s2[,1],r_cells_natural_s2[,5],r_cells_scenario1_s2[,5],r_cells_scenario2_s2[,5],
                               r_cells_scenario3_s2[,5]))
colnames(r_cells_s2) <- c("cell","natural","Alt. E","Alt. B","Alt. A")

r_prob_natural_s2 <- read.csv(file ="/Users/lohebert/GilaBN/output/natural_s2_recruit_prob.csv", header=TRUE, sep=",")
r_prob_scenario1_s2 <-read.csv(file ="/Users/lohebert/GilaBN/output/scenario1_s2_recruit_prob.csv", header=TRUE, sep=",")
r_prob_scenario2_s2 <- read.csv(file ="/Users/lohebert/GilaBN/output/scenario2_s2_recruit_prob.csv", header=TRUE, sep=",")
r_prob_scenario3_s2 <- read.csv(file ="/Users/lohebert/GilaBN/output/scenario3_s2_recruit_prob.csv", header=TRUE, sep=",")

r_prob_s2 <- cbind(r_prob_natural_s2[,2], r_prob_scenario1_s2[,2],r_prob_scenario2_s2[,2],r_prob_scenario3_s2[,2])
colnames(r_prob_s2)<- c("natural","Alt. E","Alt. B","Alt. A")  
rownames(r_prob_s2)<- c("bin_1", "bin_2", "bin_3", "bin_4", "bin_5","bin_6","bin_7","bin_8","bin_9", "bin_10", "bin_11", "bin_12", "bin_13","bin_14","bin_15")

```

Load and plot TEST
```{r}

r_cells_scenario1_s2 <-read.csv(file ="/Users/lohebert/GilaBN/output/scenario1_s2_q_cells_recruit.csv", header=TRUE, sep=",", 
                                col.names = c("cell","flow","bin","recruit_q","recruit"))
r_cells_scenario2_s2 <- read.csv(file ="/Users/lohebert/GilaBN/output/scenario2_s2_q_cells_recruit.csv", header=TRUE, sep=",", 
                                 col.names = c("cell","flow","bin","recruit_q","recruit"))
r_cells_scenario3_s2 <- read.csv(file ="/Users/lohebert/GilaBN/output/scenario3_s2_q_cells_recruit.csv", header=TRUE, sep=",", 
                                 col.names = c("cell","flow","bin","recruit_q","recruit"))

r_cells_s2 <- data.frame(cbind(r_cells_scenario1_s2[,1],r_cells_scenario1_s2[,5],r_cells_scenario2_s2[,5],
                               r_cells_scenario3_s2[,5]))
colnames(r_cells_s2) <- c("cell","Alt. E","Alt. B","Alt. A")

diff3_s2 <- cbind(r_cells_s2[1],(r_cells_s2[,2:3]-r_cells_s2[,4])*100)

colnames(diff3_s2) <- c("cell", "diffA_AltE","diffA_AltB" )

boxplot(diff3_s2[,2:3],range = 1.5, main="Site 2 - Middle- Diff. from Alt.A", ylim=c(-6, .5),names = c("Alt.E","Alt.B"),xlab = "scenarios",
        ylab= (expression(paste(Delta," posterior probability of recruitment (%)"))))

hydrodiff3_s2 <- cbind(hydro[1],(hydro[,3:4]-hydro[,5]))
colnames(hydrodiff3_s2) <- c("date", "diffA_AltE","diffA_AltB" )

boxplot(hydrodiff3_s2[,2:3],range = 1.5, main="Site 2 - Middle- Diff. from Alt.A", ylim=c(-200, 10),names = c("Alt.E","Alt.B"),xlab = "scenarios",
        ylab= (expression(paste(Delta," instream flow (cfs)"))))


```

<!-- Load scenario results for site 3 (South) to be used for analysis. -->
<!-- ```{r} -->
<!-- r_cells_natural_s3 <- read.csv(file ="/Users/lohebert/GilaBN/output/natural_s3_q_cells_recruit.csv", header=TRUE, sep=",",  -->
<!--                                col.names =  c("cell","flow","bin","recruit_q","recruit")) -->
<!-- r_cells_scenario1_s3 <-read.csv(file ="/Users/lohebert/GilaBN/output/scenario1_s3_q_cells_recruit.csv", header=TRUE, sep=",",  -->
<!--                                 col.names = c("cell","flow","bin","recruit_q","recruit")) -->
<!-- r_cells_scenario2_s3 <- read.csv(file ="/Users/lohebert/GilaBN/output/scenario2_s3_q_cells_recruit.csv", header=TRUE, sep=",",  -->
<!--                                  col.names = c("cell","flow","bin","recruit_q","recruit")) -->
<!-- r_cells_scenario3_s3 <- read.csv(file ="/Users/lohebert/GilaBN/output/scenario3_s3_q_cells_recruit.csv", header=TRUE, sep=",",  -->
<!--                                  col.names = c("cell","flow","bin","recruit_q","recruit")) -->

<!-- r_cells_s3 <- data.frame(cbind(r_cells_natural_s3[,1],r_cells_natural_s3[,5], r_cells_scenario1_s3[,5],r_cells_scenario2_s3[,5], -->
<!--                                r_cells_scenario3_s3[,5])) -->
<!-- colnames(r_cells_s3) <- c("cell","natural","Alt. E","Alt. B","Alt. A") -->

<!-- r_prob_natural_s3 <- read.csv(file ="/Users/lohebert/GilaBN/output/natural_s3_recruit_prob.csv", header=TRUE, sep=",") -->
<!-- r_prob_scenario1_s3 <-read.csv(file ="/Users/lohebert/GilaBN/output/scenario1_s3_recruit_prob.csv", header=TRUE, sep=",") -->
<!-- r_prob_scenario2_s3 <- read.csv(file ="/Users/lohebert/GilaBN/output/scenario2_s3_recruit_prob.csv", header=TRUE, sep=",") -->
<!-- r_prob_scenario3_s3 <- read.csv(file ="/Users/lohebert/GilaBN/output/scenario3_s3_recruit_prob.csv", header=TRUE, sep=",") -->

<!-- r_prob_s3 <- cbind(r_prob_natural_s3[,2], r_prob_scenario1_s3[,2],r_prob_scenario2_s3[,2],r_prob_scenario3_s3[,2]) -->
<!-- colnames(r_prob_s3)<- c("natural","Alt. E","Alt. B","Alt. A")   -->
<!-- rownames(r_prob_s3)<- c("bin_1", "bin_2", "bin_3", "bin_4", "bin_5","bin_6","bin_7","bin_8","bin_9", "bin_10", "bin_11", "bin_12", "bin_13","bin_14","bin_15") -->
<!-- ``` -->


#Calculate the differences between current conditions (scenario 1) and scenario 2 and scenario 3# (not working)
```{r}


#diffN_s2 <- cbind(r_cells_s2[1],(r_cells_s2[,2:3]-r_cells_s2[,4]))
diff3_s2 <- cbind(r_cells_s2[1],(r_cells_s2[,2:3]-r_cells_s2[,4])*100)
diff3_s2 <- cbind(r_cells_s2[1],(r_cells_s2[,3]-r_cells_s2[,4])*100,(r_cells_s2[,3]-r_cells_s2[,4])*100)


#colnames(diffN_s2) <- c("cell","diffN_AltE","diffN_AltB","diffN_AltA" )
colnames(diff3_s2) <- c("cell", "diffA_AltE","diffA_AltB" )

boxplot(diff3_s2[,2:3],range = 1.5, main="Site 2 - Middle- Diff. from Alt.A", ylim=c(-0.1, 0.1),names = c("Alt.E","Alt.B"),xlab = "scenarios",
        ylab= (expression(paste(Delta," posterior probability of recruitment (%)"))))
    
```


# Calculate the mean and standard devation of probability differences between existing conditions and each scenario
summarize default boxplot

The solid bar is the median and the whiskers are the first and third quartile
discussion/explanation of the whiskers in r:https://www.r-bloggers.com/whisker-of-boxplot/

upper whisker = min(max(x), Q_3 + 1.5 * IQR) 
lower whisker = max(min(x), Q_1 – 1.5 * IQR)


Box Plot for all scenarios for site 2.
```{r}

par(mfrow=c(1,1))

boxplot(r_cells_s2[,2:4],range = 1.5, main="Site 2 - Middle", ylim=c(0, 0.6),names = c("Alt.E","Alt.B","Alt.A"),xlab = "scenarios",
        ylab="posterior probability of recruitment")
par(mfrow=c(1,1))
```

Box Plot of differences for all scenarios for site 2.
```{r}

par(mfrow=c(1,2))

# boxplot(diffN_s2[,2:4],range = 1.5, main="Site 2 - Middle- Diff. from Natural", ylim=c(0, 0.7),names = c("Alt.E","Alt.B","Alt.A"),xlab = "scenarios",
#         ylab="change in posterior probability of recruitment")

boxplot(diff3_s2[,2:3],range = 1.5, main="Site 2 - Middle- Diff. from Alt.A", ylim=c(-0.5, 5),names = c("Alt.E","Alt.B"),xlab = "scenarios",
        ylab= (expression(paste(Delta," posterior probability of recruitment (%)"))))

par(mfrow=c(1,1))
```



```{r}
par(mfrow=c(1,3))

boxplot(r_cells_s1[,3:5],range = 1.5, main="Site 1 - North", ylim=c(0, 0.7),names = c("1","2","3"),
        ylab="posterior probability of recruitment")
boxplot(r_cells_s2[,3:5],range = 1.5, main="Site 2 - Middle", ylim=c(0, 0.7),names = c("1","2","3"),xlab = "scenarios")
boxplot(r_cells_s3[,3:5],range = 1.5, main="Site 3 - South", ylim=c(0, 0.7),names = c("1","2","3")) 

ggplot(data = r_cells_s1[,3:5]) + 
  geom_boxplot(aes(x=1,y=current_operations))+
  geom_boxplot(aes(x=2,y=combined_operations))+
  geom_boxplot(aes(x=3,y=combined_operations_AWSA))+
  xlab("scenarios")+
  ylab("posterior probability of recruitment")+
  ggtitle("Site 1 - North")

ggplot(data = r_cells_s2[,3:5]) + 
  geom_boxplot(aes(x=1,y=current_operations))+
  geom_boxplot(aes(x=2,y=combined_operations))+
  geom_boxplot(aes(x=3,y=combined_operations_AWSA))+
  xlab("scenarios")+
  ylab("posterior probability of recruitment")+
  ggtitle("Site 2 - Middle")

ggplot(data = r_cells_s3[,3:5]) + 
  geom_boxplot(aes(x=1,y=current_operations))+
  geom_boxplot(aes(x=2,y=combined_operations))+
  geom_boxplot(aes(x=3,y=combined_operations_AWSA))+
  xlab("scenarios")+
  ylab("prosterior probability of recruitment")+
  ggtitle("Site 3 - South")

par(mfrow=c(1,1))
```

Boxplot Diff
```{r}

par(mfrow=c(1,3))

boxplot(r_cells_s1[,6:7],range = 1.5, main="Site 1 - North", ylim=c(0, 0.7),names = c("diff2","diff3"),
        ylab="posterior probability of recruitment") 
boxplot(r_cells_s2[,6:7],range = 1.5, main="Site 2 - Middle", ylim=c(0, 0.7),names = c("diff2","diff3"),xlab = "scenarios")
boxplot(r_cells_s3[,6:7],range = 1.5, main="Site 3 - South", ylim=c(0, 0.7),names = c("diff2","diff3") )

par(mfrow=c(1,1))
```

# Plot mean and standard devation of changes in probabilities for each scenario

```{r}


```

# Calculate the average number of events during each time state

```{r}


```

# Calculate difference in events each season for each scenario

```{r}


```


# Plot the number of events in each month

```{r}


```

#Plot the mean probability at each site

```{r}
ncell(r_cells_s1)

hist(r_cells_natural_s1, 
     maxpixels=ncell(r_cells_natural_s1[,2]),
     main="Distribution of DSM Values\n All Pixel Values Included\n NEON Harvard Forest Field Site",
     xlab="DSM Elevation Value (m)",
     ylab="Frequency",
     col="yellow")
```


# Plot the mean probability at each site
Load cell coordinates for site 1,2,and 3 (North, Middle, and South). Pair the recruitment probability with the coordinates and plot.
*middle coordinate file is wrong
```{r}
north_coord <- read.csv("data/SiteN_xyz.txt" , header=FALSE, sep= "", col.names =  c("cell","x","y","z"))
middle_coord <- read.csv("data/SiteM_xyz.txt" , header=FALSE, sep="", col.names =  c("cell","x","y","z"))
south_coord <- read.csv("data/SiteS_xyz.txt" , header=FALSE, sep="", col.names =  c("cell","x","y","z"))

north_rloc <- merge(r_cells_s1,north_coord,by = "cell")
middle_rloc <- merge(r_cells_s2,middle_coord,by = "cell")
south_rloc <- merge(r_cells_s3, south_coord,by = "cell")

write.csv(north_rloc,file = "output/north_rloc.csv")
write.csv(middle_rloc,file = "output/middle_rloc.csv")
write.csv(south_rloc,file = "output/south_rloc.csv")

panel.mapplot(north_rloc$x,north_rloc$y,map,colramp, exact=TRUE)

panel.mapplot(x, y, map, breaks, colramp, exact = FALSE, lwd = 0.5, ...)



```




The code chunk below adds all the .tif raster files for the LiDAR and displays the crs, extent, and resolution.
```{r}
north_DEM <- raster("LiDAR_Processed/GilaNGnd.tif")
middle_DEM <- raster("LiDAR_Processed/GilaMGnd.tif")
south_DEM <- raster("LiDAR_Processed/GilaSGnd.tif")

crs(north_DEM)
extent(north_DEM)
yres(north_DEM)
xres(north_DEM)
maxValue(north_DEM)
minValue(north_DEM)

crs(middle_DEM)
extent(middle_DEM)
yres(middle_DEM)
xres(middle_DEM)
maxValue(middle_DEM)
minValue(middle_DEM)

crs(south_DEM)
extent(south_DEM)
yres(south_DEM)
xres(south_DEM)
maxValue(south_DEM)
minValue(south_DEM)

```

Plot posterior recruitment potential

```{r}
#rasterFromXYZ()

north_plot <-ggplot() + geom_raster(data = north_rloc, aes(x=north_rloc$x, y = north_rloc$x, fill=north_rloc$combined_operations)) 

north_plot

```


```{r}
north_plot <- plot(north_rloc$x,north_rloc$y, type="p", col = c("blue","lightblue","green","yellow", "red"), pch=20, cex = 0.05, asp=1)
middle_plot <-plot(middle_rloc$x,middle_rloc$y, type="p", col = c("blue","lightblue","green","yellow", "red"), pch=20, cex = 0.05, asp=1 ) 
south_plot <- plot(south_rloc$x,south_rloc$y, type="p", col = c("blue","lightblue","green","yellow", "red"), pch=20, cex = 0.05,asp=1 ) 

north_plot <-ggplot() + geom_raster(data = test_data, aes(x=Lon, y = Lat, fill=Value)) + 
             coord_fixed(ratio = 1) +
             scale_fill_viridis(direction = -1) +
             theme_bw()
```


#Plot diff#
```{r}
north_plot <- plot(north_rloc$x,north_rloc$y, type="p", col = c("blue","lightblue","green","yellow", "red"), pch=20, cex = 0.05, asp=1)
middle_plot <-plot(middle_rloc$x,middle_rloc$y, type="p", col = c("blue","lightblue","green","yellow", "red"), pch=20, cex = 0.05, asp=1 ) 
south_plot <- plot(south_rloc$x,south_rloc$y, type="p", col = c("blue","lightblue","green","yellow", "red"), pch=20, cex = 0.05,asp=1 ) 

```


Plots/maps the sites (North, Middle, and South) based on 2018 LiDAR and maps the recruitment potential.
```{r}
plot(north_DEM, col=greyPalette(25), main="Upper Gila North")
```

Plots/maps the sites (North, Middle, and South) based on 2018 LiDAR and maps the recruitment potential.
```{r}
plot(north, col=greyPalette(25), main="Upper Gila North")
plot(middle, col=greyPalette(25), main="Upper Gila Middle")
plot(south, col=greyPalette(25), main="Upper Gila South")
```

Plots/maps the sites (North, Middle, and South) based on 2018 LiDAR and maps the recruitment potential.
```{r}
plot(north, col=greyPalette(25), main="Upper Gila North")
plot(middle, col=greyPalette(25), main="Upper Gila Middle")
plot(south, col=greyPalette(25), main="Upper Gila South")
```
Plots that shows the topography of the sites and profiles. 
```{r}
cols <- colorRampPalette(brewer.pal(9,"BrBG"))

levelplot(north_DEM, col.regions=cols, main="2018 LiDAR \nUpper Gila North")
levelplot(middle_DEM, col.regions=cols, main="2018 LiDAR \nUpper Gila Middle")
levelplot(south_DEM, col.regions=cols, main="2018 LiDAR \nUpper Gila South")
```

##References##
