==> 5-gila-optimization.Rmd <==
---
title: Upper Gila Watershed Bayesian Network Optimization
author: Lauren Victoria Jaramillo, UNM (alwayslo@unm.edu) & Jonathan Carter Hebert
date: 6/26/2019
update: 8/21/2019
output: html_notebook
---

This R notebook is used to used to optimize the BN model.

###Optimization###
```{r}
# rm(list=ls())

library("bnlearn")
library("readr")
library("Rgraphviz")
library("xts")
library("prob")
library("TTR")
library("bigmemory")
library("zoo")
library("raster")
library("sp")
library("rgdal")
library("rasterVis")
library("ggplot2")
library("rgdal")
library("methods")
require("BiocGenerics")
require("graph")
library("maps")
library("RColorBrewer")
library("stats")
library("stringi")
library ("hydroTSM")

options("scipen" = 100, "digits"=4)

getwd()
```

Load and plot all flow scenarios modeled.
```{r}

hydro <- read.csv("data/scenarios_revised.csv", skip=0,header = TRUE)[ ,1:4]
hydro <- setNames(hydro, c("date", "Alt_E","Alt_B","Alt_A"))
hydro$date<-as.Date(hydro$date, format="%m/%d/%Y")
hydro <- cbind(hydro[1],usgs_hydro[4],hydro[2:4])
hydro <- setNames(hydro, c("date","natural", "Alt_E","Alt_B","Alt_A"))
hydro$julian <- format(hydro$date, "%j")
hydro$julian <-stri_replace_all_regex(hydro$julian, "\\b0*(\\d+)\\b", "$1")
hydro$month <- format(hydro$date,"%m")
```

Plots the hydrologic data that is actually used in the model which is governed by the lowest minimum innundation value.

Flow Duration Curve
```{r}

fdc(hydro[3:5], lQ.thr=0.7, hQ.thr=0.2, plot=TRUE, log="y",
     main= "Flow Duration Curve", xlab="% Time flow equalled or exceeded",
     ylab="Q, [ft3/s]",  yat=c(1,10,50,100,250,500,750,1000), xat=c(0.01, 0.025, 0.05),
     col=c("blue","green","orange"), pch=1:ncol(hydro), lwd=rep(1, ncol(hydro)),
     lty=1:ncol(hydro), cex=0.4, cex.axis=1.2, cex.lab=1.2, leg.txt=NULL,
     leg.cex=1, leg.pos="topright", verbose=TRUE, thr.shw=TRUE, new=TRUE)

```

Plots the mean daily flows for all four scenarios (natural,Alt_E,Alt_B,Alt_A).
```{r}
meanJulian <- data.frame()

for (i in 1:length(unique(hydro$julian))) 
{ julian_i <- colMeans(hydro[hydro$julian== i,2:5])
  meanJulian <- rbind(meanJulian, julian_i)}

meanJulian <- cbind( 1:length(unique(hydro$julian)),meanJulian)
meanJulian<-setNames(meanJulian, c("julian","natural", "Alt_E","Alt_B","Alt_A"))
meanJulian<-na.omit(meanJulian)

ggplot(meanJulian, aes(x=julian))+ 
  geom_line(aes(y = Alt_E, color = "Alt. E"),linetype = 1)+
  geom_line(aes(y = Alt_B, color = "Alt. B"),linetype = 1)+
  geom_line(aes(y = Alt_A, color = "Alt. A"),linetype = 1)+
  geom_vline(xintercept = 92, linetype="dotted",  color = "black", size=0.5)+
  geom_vline(xintercept = 274, linetype="dotted",  color = "black", size=0.5)+
  xlab("Julian day")+
  ylab("flow (cfs)")+
  ggtitle("Upper Gila Mean Daily Flow\n April to September 1936 - 2017")+
  theme(plot.title = element_text(hjust = 0.5))+
  scale_color_manual(values=c("blue","green","orange"))+
  labs(color="scenario")
  

#geom_hline(yintercept = c(1000, 1500, 2000, 2500, 3000, 3500, 4000))
```

```{r}
sigma1Julian <- data.frame()

for (i in 1:length(unique(hydro$julian))) 
{ julian_i <- colMeans(hydro[hydro$julian== i,2:5])+ colStdevs(hydro[hydro$julian== i,2:5])
  sigma1Julian <- rbind(sigma1Julian, julian_i)}

sigma1Julian <- cbind( 1:length(unique(hydro$julian)),sigma1Julian)
sigma1Julian<-setNames(sigma1Julian, c("julian","natural", "Alt_E","Alt_B","Alt_A"))
sigma1Julian<-na.omit(sigma1Julian)

ggplot(sigma1Julian, aes(x=julian))+ 
  geom_line(aes(y = natural, color = "natural"),linetype = 1)+
  geom_line(aes(y = Alt_E, color = "Alt. E"),linetype = 1)+
  geom_line(aes(y = Alt_B, color = "Alt. B"),linetype = 2)+
  geom_line(aes(y = Alt_A, color = "Alt. A"),linetype = 4)+
  xlab("Julian day")+
  ylab("flow (cfs)")+
  ggtitle(expression(paste("Upper Gila 1", sigma, " Above Mean Daily Flow\n April to September 1936 - 2017")))+
  theme(plot.title = element_text(hjust = 0.5))+
  labs(color="scenario")+
  geom_vline(xintercept = 92, linetype="dotted",  color = "black", size=0.5)+
  geom_vline(xintercept = 274, linetype="dotted",  color = "black", size=0.5)+
  geom_hline(yintercept = c(1000, 1500, 2000, 2500, 3000, 3500, 4000))

```


```{r}
sigma2Julian <- data.frame()

for (i in 1:length(unique(hydro$julian))) 
{ julian_i <- colMeans(hydro[hydro$julian== i,2:5])+ 2*colStdevs(hydro[hydro$julian== i,2:5])
  sigma2Julian <- rbind(sigma2Julian, julian_i)}

sigma2Julian <- cbind( 1:length(unique(hydro$julian)),sigma2Julian)
sigma2Julian<-setNames(sigma2Julian, c("julian","natural", "Alt_E","Alt_B","Alt_A"))
sigma2Julian<-na.omit(sigma2Julian)


ggplot(sigma2Julian, aes(x=julian))+ 
  geom_line(aes(y = natural, color = "natural"),linetype = 1)+
  geom_line(aes(y = Alt_E, color = "Alt. E"),linetype = 1)+
  geom_line(aes(y = Alt_B, color = "Alt. B"),linetype = 2)+
  geom_line(aes(y = Alt_A, color = "Alt. A"),linetype = 4)+
  xlab("Julian day")+
  ylab("flow (cfs)")+
  ggtitle(expression(paste("Upper Gila 2", sigma, " Above Mean Daily Flow\n April to September 1936 - 2017")))+
  theme(plot.title = element_text(hjust = 0.5))+
  geom_vline(xintercept = 92, linetype="dotted",  color = "black", size=0.5)+
  geom_vline(xintercept = 274, linetype="dotted",  color = "black", size=0.5)+
  labs(color="scenario")+
  geom_hline(yintercept = c(1000, 1500, 2000, 2500, 3000, 3500, 4000))



```

##Histogram##

Histogram of flows for each scenario (scenario 1, scenario 2, and scenario 3) for the mean, sigma1, and sigma2.
mar=c(bottom, left, top, right)
oma = c(bottom, left, top, right) 
```{r}

par(mfrow=c(4,3), mai = c(.25,.25,.25,.25),mar=c(4,4,.75,.5),oma = c(1,1,1,1) )

hist(meanJulian$Alt_E,
     main =  expression(paste(bar(x))),
     xlab = NULL,
     col = "blue")
hist(meanJulian$Alt_B,
     main = expression(paste(bar(x))), 
     xlab = NULL,
     ylab = NULL,
     col = "green")
hist(meanJulian$Alt_A,
     main = expression(paste(bar(x))), 
     xlab = NULL,
     ylab = NULL,
     col = "orange")

hist(sigma1Julian$Alt_E,
     main = expression(paste( sigma, "=1")), 
     xlab = NULL,
     col = "blue")
hist(sigma1Julian$Alt_B,
     main = expression(paste( sigma, "=1")), 
     xlab = NULL,
     ylab = NULL,
     col = "green")
hist(sigma1Julian$Alt_A,
     main = expression(paste( sigma, "=1")), 
     xlab = NULL,
     ylab = NULL,
     col = "orange")

hist(sigma2Julian$Alt_E,
     main = expression(paste( sigma, "=2")), 
     xlab = "flow (cfs)",
     col = "blue")
hist(sigma2Julian$Alt_B,
     main = expression(paste( sigma, "=2")), 
     xlab = "flow (cfs)",
     ylab = NULL,
     col = "green")
hist(sigma2Julian$Alt_A,
     main = expression(paste( sigma, "=2")), 
     xlab = "flow (cfs)",
     ylab = NULL,
     col = "orange")

plot(3, type = "n", axes=FALSE, xlab="", ylab="")
legend("left",
       legend=c("Alt. E", "Alt. B", "Alt. A"), 
       col = c("blue","green","orange"),
       xpd=NA, 
       pch = c(15, 15, 15))

par(mfrow=c(1,1))

```

Histogram of Differences 

Histogram of differences for each scenario (scenario 1, scenario 2, and scenario 3) for the mean, sigma1, and sigma2.
mar=c(bottom, left, top, right)
oma = c(bottom, left, top, right) 
```{r}

par(mfrow=c(4,3), mai = c(.25,.25,.25,.25),mar=c(4,4,.75,.5),oma = c(1,1,1,1) )

hist(meanJulian$natural - meanJulian$Alt_E,
     main =  expression(paste(bar(x))),
     xlab = "diversion (cfs)",
     col = "blue")
hist(meanJulian$natural - meanJulian$Alt_B,
     main = expression(paste(bar(x))), 
     xlab = "diversion (cfs)",
     ylab = NULL,
     col = "green")
hist(meanJulian$natural - meanJulian$Alt_A,
     main = expression(paste(bar(x))), 
     xlab = "diversion (cfs)",
     ylab = NULL,
     col = "orange")

hist(sigma1Julian$natural - sigma1Julian$Alt_E,
     main = expression(sigma),
     xlab = "diversion (cfs)",
     col = "blue")
hist(sigma1Julian$natural - sigma1Julian$Alt_B,
     main = expression(sigma), 
     xlab = "diversion (cfs)",
     ylab = NULL,
     col = "green")
hist(sigma1Julian$natural - sigma1Julian$Alt_A,
     main = expression(sigma), 
     xlab = "diversion (cfs)",
     ylab = NULL,
     col = "orange")

hist(sigma2Julian$natural - sigma2Julian$Alt_E,
     main = expression(paste(sigma,"=2")),
     xlab = "diversion (cfs)",
     col = "blue")
hist(sigma2Julian$natural - sigma2Julian$Alt_B,
     main = expression(paste(sigma,"=2")), 
     xlab = "diversion (cfs)",
     ylab = NULL,
     col = "green")
hist(sigma2Julian$natural - sigma2Julian$Alt_A,
     main = expression(paste(sigma,"=2")), 
     xlab = "diversion (cfs)",
     ylab = NULL,
     col = "orange")

plot(3, type = "n", axes=FALSE, xlab="", ylab="")
legend("left",
       legend=c("Alt. E", "Alt. B", "Alt. A"), 
       col = c("blue","green","orange"),
       xpd=NA, 
       pch = c(15, 15, 15))

par(mfrow=c(1,1))

```


Difference Plot

Difference Signal

This plots the difference between each scenario and the 'natural' scenario. 
```{r}
ggplot(meanJulian, aes(x=julian))+ 
  geom_line(aes(y = natural - Alt_E, color = "Alt. E"),linetype = 1)+
  geom_line(aes(y = natural - Alt_B, color = "Alt. B"),linetype = 2)+
  geom_line(aes(y = natural - Alt_A, color = "Alt. A"),linetype = 4)+
  xlab("Julian day")+
  ylab("flow (cfs)")+
  ggtitle("Upper Gila Difference in Mean Daily Flow\n April to September 1936 - 2017")+
  theme(plot.title = element_text(hjust = 0.5))+
  labs(color="scenario")
  #geom_hline(yintercept = c(1000, 1500, 2000, 2500, 3000, 3500, 4000))
```





###FFT (IGNORE - NOT COMPLETE)### 

Computes the Discrete Fourier Transform (DFT) of an array with a fast algorithm, the “Fast Fourier Transform” (FFT) and plots the results for each scenario.
```{r}

diff_scenario1 <- abs(fft(hydro$natural - hydro$Alt_E, inverse= FALSE))
diff_scenario2 <- abs(fft(hydro$natural - hydro$Alt_B, inverse= FALSE))
diff_scenario3 <- abs(fft(hydro$natural - hydro$Alt_A, inverse= FALSE))

plot(diff_scenario1)
plot(diff_scenario2)
plot(diff_scenario3)

```

Computes the Discrete Fourier Transform (DFT) of an array (annualized mean flow) with a fast algorithm, the “Fast Fourier Transform” (FFT) and plots the results for each scenario. meanJulian
```{r}

mj_diff_scenario1 <- fft(meanJulian$natural - meanJulian$Alt_E, inverse= FALSE)
mj_diff_scenario2 <- fft(meanJulian$natural - meanJulian$Alt_B, inverse= FALSE)
mj_diff_scenario3 <- fft(meanJulian$natural - meanJulian$Alt_A, inverse= FALSE)

plot(mj_diff_scenario1)
plot(mj_diff_scenario2)
plot(mj_diff_scenario3)
```

frequence


```{r}
library(dplyr)
library(useful)

Xk<-data.frame(cart2pol(Re(mj_diff_scenario1), Im(mj_diff_scenario1),degrees = FALSE))
Xk$n = 1:nrow(Xk)


Xk$k <- Xk$theta/(2*pi)*(366/Xk$n)

mj_diff_scenario1_inv <- fft(mj_diff_scenario1, inverse= TRUE)/366
xn <-data.frame(cart2pol(Re(mj_diff_scenario1_inv), Im(mj_diff_scenario1_inv),degrees = FALSE))
xn$n = 1:nrow(xn)
xn$k <- xn$theta/(2*pi)*(366/xn$n)

plot(xn$x,xn$k)

```

###bandwidth###


#References#

https://stat.ethz.ch/R-manual/R-devel/library/stats/html/fft.html

https://www.rdocumentation.org/packages/stats/versions/3.6.0/topics/bandwidth

http://www.di.fc.ul.pt/~jpn/r/fourier/fourier.html

Vogel, R., and N. M. Fennessey (1994), Flow duration curves I: A new interpretation and confidence intervals, ASCE, Journal of Water Resources Planning and Management, 120(4). 

Vogel, R., and N. Fennessey (1995), Flow duration curves II: A review of applications in water resources planning, Water Resources Bulletin, 31(6), 1029-1039, doi:10.1111/j.1752-1688.1995.tb03419.x. 

Yilmaz, K. K., H. V. Gupta, and T. Wagener (2008), A process-based diagnostic approach to model evaluation: Application to the NWS distributed hydrologic model, Water Resour. Res., 44, W09417, doi:10.1029/2007WR006716.

https://www.rforge.net/doc/packages/hydroTSM/fdc.html
